% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_process_functions.R
\name{anomG3est}
\alias{anomG3est}
\title{Perfrom anomaly G3est on a \code{\link[spatstat]{pp3}} object.}
\usage{
anomG3est(pattern, result, correction = "rs")
}
\arguments{
\item{pattern}{The \code{\link[spatstat]{pp3}} object to analyze.}

\item{result}{List returned from \code{\link{pG3est}} with \code{anom =
TRUE}.}

\item{correction}{See \code{\link{pG3est}}. "rs", "km", "Hanisch", or "all".}
}
\value{
Returns data fram containing r values and associated anomaly G3est
  values.
}
\description{
See \code{\link[spatstat]{G3est}}. Performs the anomaly G3est on a set of
point cloud data. This means subtracting the 50th percentile from the
results. This centers the curve around zero. Used for comparing data to
envelopes from \code{\link{pG3est}} where \code{anom = TRUE}. Will subtract
the same values used in the pG3est test that is being compared to.
}
