% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_localK3.R
\name{anomlocalK3est}
\alias{anomlocalK3est}
\title{Perform localK3est with 50th percentile of a RRL subtracted off.}
\usage{
anomlocalK3est(X, toSub, rmax, nrval)
}
\arguments{
\item{X}{The \code{\link[spatstat]{pp3}} object to be tested.}

\item{toSub}{The vector of values to subtract from the square root of the
results of the K function applied to X.}

\item{rmax}{See \code{\link[spatstat]{K3est}}.}

\item{nrval}{See \code{\link[spatstat]{K3est}}.}
}
\value{
Date frame with columns of the anomaly K test for each point in the
  pattern.
}
\description{
Similar to \code{\link{anomK3est}}, but returns the anomaly K3est for each
point in the pattern.
}
