% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_summarize.R
\name{average_relabelings}
\alias{average_relabelings}
\title{Functionto take all relabelings and return averages and envelope values}
\usage{
average_relabelings(
  relabelings,
  envelope.value = 0.95,
  funcs = c("K", "G"),
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{relabelings}{output of  \code{\link[rapt]{relabel_summarize}} function}

\item{envelope.value}{size of envelope to compute.  Should be decimal (e.g. 0.95 = 95\%)}
}
\description{
Functionto take all relabelings and return averages and envelope values
}
