% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{bcc.gen-deprecated}
\alias{bcc.gen-deprecated}
\title{Generate body centerd cubic (BCC) lattice of points in 3D.}
\arguments{
\item{npoint}{Approximate number of points to have in the pattern.}

\item{win}{A \code{\link[spatstat.geom]{box3}} object containing the window for
the final pattern.}
}
\value{
A \code{\link[spatstat.geom]{pp3}} object with the lattice points.
}
\description{
Generates a BCC point pattern with a specified window size and number of
points desired in the pattern.
}
\seealso{
\code{\link{rapt-deprecated}}
}
\keyword{internal}
