% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{bdist.points.pp3}
\alias{bdist.points.pp3}
\title{Distance to Boundary of Domain}
\usage{
\method{bdist.points}{pp3}(X)
}
\arguments{
\item{X}{The point pattern for analysis. A \code{\link[spatstat.geom]{pp3}}
object.}
}
\value{
An object containing the shortest distance to the boundary for each
point in the pattern X.
}
\description{
Finds the smallest distance to a boundary for each point in a point pattern.
}
\seealso{
\code{\link[spatstat]{bdist.points}}
}
