% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{bdist.points.ppp}
\alias{bdist.points.ppp}
\title{Distance to Boundary of Window}
\usage{
\method{bdist.points}{ppp}(X)
}
\description{
Distance to Boundary of Window
}
\seealso{
\code{\link[spatstat]{bdist.points}}
}
