% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-extend.R
\name{bdist.points3.multi}
\alias{bdist.points3.multi}
\title{Returns the shortest distances to boundaries in the x, y, and z directions
separately.}
\usage{
bdist.points3.multi(X)
}
\arguments{
\item{X}{The point pattern for analysis. A \code{\link[spatstat]{pp3}}
object.}
}
\value{
A data.frame containing the shortest distance to the closest three
  boundaries for each point in the pattern X.
}
\description{
Returns the shortest distances to boundaries in the x, y, and z directions
separately.
}
\seealso{
\code{\link{bdist.points3}}
}
