% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_process_functions.R
\name{bdist.points3}
\alias{bdist.points3}
\title{Helper function for border correction \code{\link{bK3est}}.}
\usage{
bdist.points3(X)
}
\arguments{
\item{X}{The point pattern for analysis. A \code{\link[spatstat]{pp3}} object.}
}
\value{
An object containing the shortest distance to the boundary for each
  point in the pattern X.
}
\description{
Finds the smallest distance to a boundary for each point in a point pattern.
}
