% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_summarize.R
\name{calc_summary_funcs}
\alias{calc_summary_funcs}
\title{function to calculate all summary functions on point pattern}
\usage{
calc_summary_funcs(
  pattern,
  funcs = c("K", "G", "F", "GXGH"),
  dopant_formula,
  host_formula,
  k = 1,
  maxKr = 10,
  nKr = 200,
  maxGr = 5,
  nGr = 1000,
  maxGXGHr = 3,
  maxGXHGr = 8,
  nGXr = 1000,
  vside = 0.3,
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{pattern}{point pattern of type ppp or pp3}

\item{funcs}{vector of summary functions to calculate}

\item{dopant_formula}{mark of points from which to calculate summary functions}

\item{host_formula}{marks of points to not use in summary functions (except cross type functions)}

\item{K_cor}{edge correction(s) to be used for K function}

\item{G_cor}{edge correction(s) to be used for G function}

\item{F_cor}{edge correction(s) to be used for F function}

\item{GXGH_cor}{edge correction(s) to be used for GXGH function}

\item{GXHG_cor}{edge correction(s) to be used for GXHG function}

\item{...}{maxKr, nKr, maxGr, nGr, maxGXGHr, maxGXHGr, nGXr}
}
\description{
Cacluate that summary functions included in \code{funcs} on \code{pattern}, a point pattern
of class ppp or pp3.
}
