% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{cformCount}
\alias{cformCount}
\title{Count the number of hits for each entry within a \code{RNG} object}
\usage{
cformCount(pos, cform)
}
\description{
Count the number of hits for each entry within a \code{RNG} object
}
