% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{cformPat}
\alias{cformPat}
\title{Generate the isotopic positions and intensities from a \code{cform}
object.}
\usage{
cformPat(cform, charge = 1:4, thresh = 20)
}
\description{
Generate the isotopic positions and intensities from a \code{cform}
object.
}
