% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{crAdjust.new}
\alias{crAdjust.new}
\title{Helper for \code{\link{makecluster}} for \code{type} = "cr" and \code{fast} =
TRUE}
\usage{
crAdjust.new(cluster.ind, cluster.info, diff, X, Y)
}
\arguments{
\item{cluster.ind}{Vector of indices of the cluster points.}

\item{cluster.info}{Factor containing the number of points in each cluster.}

\item{diff}{The difference between the number of values needed and the number
in cluster.ind.}

\item{X}{The overlaying point pattern. A \code{\link[spatstat.geom]{pp3}} object.}

\item{Y}{The underlaying point pattern. A \code{\link[spatstat.geom]{pp3}} object.}
}
\value{
A list of [\link{1}] The updated cluster.ind vector and [\link{2}] the updated
cluster.info factor.
}
\description{
Adjustment method to get correct number of points in each cluster for the
\code{type} = "cr" cluster generation method of \code{\link{makecluster}}
function. This is a version of \code{\link{crAdjust}} meant to work with the
faster, updated verios of \code{type}="cr" \code{\link{makecluster}}.
}
