% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R, R/rapt_sim-cluster.R
\name{crAdjust}
\alias{crAdjust}
\title{Helper for \code{\link{makecluster}} for \code{type} = "cr" and \code{fast} =
FALSE}
\usage{
crAdjust(mat, diff, X, Y)

crAdjust(mat, diff, X, Y)
}
\arguments{
\item{mat}{Matrix filled with points and associated cluster index values.}

\item{diff}{The difference between the number of values needed and the number
in mat}

\item{X}{The overlaying point pattern. A \code{\link[spatstat]{pp3}} object.}

\item{Y}{The underlaying point pattern. A \code{\link[spatstat]{pp3}} object.}

\item{mat}{A matrix. The matrix filled with points and associated cluster
index values.}

\item{diff}{A numeric. The difference between the number of values needed
and the number in \code{mat}.}

\item{X}{A \code{pp3}. The over point pattern.}

\item{Y}{A \code{pp3}. The under point pattern.}
}
\value{
Updated mat matrix containing the correct number of cluster points.
}
\description{
Adjustment method to get correct number of points in each cluster for the
\code{type} = "cr" cluster generation method of \code{\link{makecluster}}
function.
}
