% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{create_training_data}
\alias{create_training_data}
\title{Feature Extraction}
\usage{
create_training_data(
  i,
  pattern,
  rrl,
  nper = 1,
  params,
  pcp = 0.5,
  tol = 0.005,
  maxKr = 10,
  nKr = 200,
  maxGr = 5,
  nGr = 1000,
  maxGXGHr = 3,
  maxGXHGr = 8,
  nGXr = 1000,
  vside = 0.3,
  jitter = 0,
  feats = c("G_max_diff", "G_max_diff_r", "G_min_diff", "G_zero_diff_r", "F_min_diff",
    "F_min_diff_F", "Tm", "Rm", "Rdm", "Rddm", "Tdm", "GXGH_min_diff", "GXGH_95diff_r",
    "GXGH_FWHM"),
  total_time_start = as.numeric(Sys.time())
)
}
\arguments{
\item{i}{an integer.  This will be the iteration number.  It determines the seed number
in the \code{\link[rapt]{clustersim}} function.}

\item{pattern}{a 3D point pattern (object of type \code{pp3}) to use as the underlying point pattern (UPP) onto
which generate clusters}

\item{rrl}{output of  \code{\link[rapt]{average_relabelings}}.  Used as expected values for summary functions}

\item{pcp}{a numeric.  Guest concentration}

\item{tol}{a numeric. 'tol' or tolerance in \code{\link[rapt]{clustersim}}}

\item{maxKr}{a numeric.  Maximum value to calculate K to in \code{\link[spatstat]{K3est}}}

\item{nKr}{a numeric.  Number of intervals at which to calculate K in  \code{\link[spatstat]{K3est}}}

\item{maxGr}{a numeric.  Maximum value to calculate G to in \code{\link[spatstat]{G3est}}}

\item{nGr}{a numeric.  Number of intervals at which to calculate G in  \code{\link[spatstat]{G3est}}}

\item{maxGXGHr}{a numeric.  Maximum value to calculate GXGH to in \code{\link[rapt]{G3cross}}}

\item{maxGXHGr}{a numeric.  Maximum value to calculate GXHG to in \code{\link[rapt]{G3cross}}}

\item{nGXr}{a numeric.  Number of intervals at which to calculate G in  \code{\link[rapt]{G3cross}}}

\item{vside}{a numeric. voxel side length for \code{\link[spatstat]{F3est}}}

\item{jitter}{a numeric.  Scale of perturbations. 'radius' value in  \code{\link[rapt]{rjitter.pp3}}}

\item{feats}{a vector.  Names of all features to be kept}

\item{total_time_stat}{a numeric. Leave as default to let it algorithm calculate iteration time}
}
\description{
function for extracting features for ML algorithm
}
\details{
Take an point pattern, simulate clustering, calculate summary functions on this new point
pattern, extract features from summary functions.
}
