% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-file.R
\name{createSpat}
\alias{createSpat}
\title{Create a \code{\link[spatstat]{pp3}} object from a POS or ATO data frame.}
\usage{
createSpat(pos, win = NULL)
}
\arguments{
\item{pos}{A POS or ATO data frame.}
}
\value{
A \code{\link[spatstat]{pp3}} with the x,y,z positions of the hits in
  the supplied POS or ATO.
}
\description{
Create a \code{\link[spatstat]{pp3}} object from a POS or ATO data frame.
}
\seealso{
\code{\link{readPOS}}, \code{\link{readATO}}
}
