% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{createTOF}
\alias{createTOF}
\title{Create a TOF Spectrum from an ATO}
\usage{
createTOF(pos, res = 0.001)
}
\description{
\code{createTOF} generates a TOF spectrum (using
\code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}) from an ATO
\code{data.frame}.
}
\seealso{
\code{\link{createSpec}}
}
