% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndist_analysis_functions.R
\name{data_frame_list_averager}
\alias{data_frame_list_averager}
\title{Average data frames}
\usage{
data_frame_list_averager(
  data,
  output = "plot",
  location = getwd(),
  image_name = "image"
)
}
\value{
Describe output
}
\description{
add description
}
\details{
add details

 @param image
 @param image2
}
