% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-vis.R
\name{detectorGIF}
\alias{detectorGIF}
\title{Create a GIF that steps through detector hits.}
\usage{
detectorGIF(ato, len = 1e+05, range = NULL, size = 7,
  name = "detector", path = "./", save.pdf = F, delay = 10, ...)
}
\description{
\code{detectorGIF} generates a GIF that shows the positions of detector hits.
}
