% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_multi.R
\name{dissocDensity}
\alias{dissocDensity}
\title{Project hits along a dissociation track}
\usage{
dissocDensity(
  sax,
  m1,
  m2,
  mp,
  dv = NULL,
  v = NULL,
  drop = FALSE,
  nbhd = "point",
  method = "max",
  renorm = TRUE
)
}
\arguments{
\item{sax}{The output of \code{\link{saxeyPlot}}}

\item{m1}{Mass to charge ratio of the first daughter ion}

\item{m2}{Mass to charge ratio of the second daughter ion}

\item{mp}{Mass to charge ratio of the parent ion}

\item{dv}{Voltage parameter step}

\item{v}{Voltage parameter series}

\item{drop}{Should repeated bins be dropped? Defaults to FALSE.}

\item{nbhd}{The neighborhood of bins to be included. One of: "point",
"rook", or "queen".}

\item{method}{A function or character string to be matched by
\code{\link{match.fun}} that returns a single value. The density
calculation method used for the neighborhood of each point.}

\item{renorm}{Should the returned densities be renormalized to sum to 1?
FALSE uses the densities in \code{sax}. Defaults to TRUE. See
\code{\link[ash]{ash2}}.}
}
\description{
Project hits along a dissociation track
}
\references{
Saxey, D. W. (2011). Correlated ion analysis and the
interpretation of atom probe mass spectra. Ultramicroscopy, 111(6), 473-479.
\url{https://doi.org/10.1016/j.ultramic.2010.11.021}.
}
\seealso{
\code{\link{saxeyPlot}}, \code{\link{dissocTrack}},
\code{\link[ash]{ash2}}
}
