% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_multi.R
\name{dissocTime}
\alias{dissocTime}
\title{Estimate the dissociation time}
\usage{
dissocTime(mp, x0, phi, v0, amu = 1.66053904e-27, el = 1.602176634e-19)
}
\description{
\code{dissocTime}
}
\references{
Saxey, D. W. (2011). Correlated ion analysis and the
interpretation of atom probe mass spectra. Ultramicroscopy, 111(6), 473-479.
\url{https://doi.org/10.1016/j.ultramic.2010.11.021}.
}
\seealso{
\code{\link{saxeyPlot}}, \code{\link{dissocDensity}}
}
