% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_multi.R
\name{dissocTrack}
\alias{dissocTrack}
\title{Parameterized dissociation track on a multiple hit}
\usage{
dissocTrack(v, m, mp)
}
\description{
\code{dissocTrack}
}
\references{
Saxey, D. W. (2011). Correlated ion analysis and the
interpretation of atom probe mass spectra. Ultramicroscopy, 111(6), 473-479.
\url{https://doi.org/10.1016/j.ultramic.2010.11.021}.
}
\seealso{
\code{\link{saxeyPlot}}, \code{\link{dissocDensity}}
}
