% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{drawRange}
\alias{drawRange}
\title{Color ranges on a MassSpectrum}
\usage{
drawRange(ms, start, end, col = "red")
}
\arguments{
\item{ms}{data.frame or
\code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}. The data on
on which the ranges should be drawn}

\item{start}{numeric. The start of the range.}

\item{end}{numeric. The end of the range}

\item{col}{The color of the range.}
}
\description{
\code{drawRange} draws a colored range on a plot
}
\seealso{
\code{\link{drawRNG}}
}
