% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{drawRNG}
\alias{drawRNG}
\title{Color ranges from an RNG on a MassSpectrum}
\usage{
drawRNG(ms, rng)
}
\description{
\code{drawRNG} draws colored ranges from an RNG (as generated by
\code{\link{readRNG}}) on a plot of a
\code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}
}
\seealso{
\code{\link{drawRange}}
}
