% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_functions.R
\name{envPlot}
\alias{envPlot}
\title{Plot envelopes of K3est test}
\usage{
envPlot(tests, percentiles = c(0.999, 0.99, 0.97), ylim = c(-3, 3),
  xlim = c(0, ceiling(max(tests[, 1]))))
}
\arguments{
\item{tests}{The return file from \code{\link{pK3est}} or the first, [[1]],
entry in the list returned by \code{\link{panomK3est}}. Contains results
from many 3D K tests.}

\item{percentiles}{Numerical vector of percentiles that you want to see the
envelopes for. Each between 0 and 1.}

\item{ylim}{Numerical vector containing the min and max values for the y axis
on the plot.}

\item{xlim}{Numerical vector containing the min and max values for the x axis
on the plot.}
}
\value{
Nothing.
}
\description{
Plot the results of envelope calculations from the \code{\link{pK3est}} or
\code{\link{panomK3est}}, with the ability to choose the percentiles for
plotting.
}
