% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{extract_features}
\alias{extract_features}
\title{Feature Extraction}
\usage{
extract_features(
  i,
  pattern,
  rrl,
  nper = 1,
  params,
  pcp = 0.5,
  tol = 0.005,
  maxKr = 10,
  nKr = 200,
  maxGr = 5,
  nGr = 1000,
  maxGXGHr = 3,
  maxGXHGr = 8,
  nGXr = 1000,
  vside = 0.3,
  feats = c("G_max_diff", "G_max_diff_r", "G_min_diff", "G_zero_diff_r", "F_min_diff",
    "F_min_diff_F", "Tm", "Rm", "Rdm", "Rddm", "Tdm", "GXGH_min_diff", "GXGH_95diff_r",
    "GXGH_FWHM"),
  total_time_start = as.numeric(Sys.time())
)
}
\description{
function for extracting features for ML algorithm. Not ready for use
}
