% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{f3features}
\alias{f3features}
\title{F features}
\usage{
f3features(rvals, fvals_new, fvals_old)
}
\arguments{
\item{rvals}{vector of radius values}

\item{fvals_new}{vector of new F3est values from which to subtract expected values}

\item{fvals_old}{vector of expected F3est values}
}
\value{
features to be used for machine learning prediction of clustering behavior.  The features are
the minimum difference between the functions (\emph{min_diff}), and the value of F (fvals_new) at
the minimum difference (\emph{min_diff_F})
}
\description{
#' @description
Features calculated from F function (F3est).
}
\details{
Takes one set of radius values and corresponding F values of two functions as input and returns
features for machine learning
}
