% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{feature_extraction}
\alias{feature_extraction}
\title{Feature Extraction}
\usage{
feature_extraction(
  i,
  pattern,
  rrl,
  pcp = 0.5,
  cr = 4,
  rho1 = 0.5,
  rho2 = 0.01,
  rb = 0.1,
  pb = 0.1,
  tol = 0.005,
  maxKr = 10,
  nKr = 200,
  maxGr = 5,
  nGr = 1000,
  maxGXGHr = 3,
  maxGXHGr = 8,
  nGXr = 1000,
  vside = 0.3,
  total_time_start = as.numeric(Sys.time())
)
}
\description{
Feature Extraction
}
