% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_version.R
\name{find_all_installed_package}
\alias{find_all_installed_package}
\title{find all the installed packages in the current system}
\usage{
find_all_installed_package()
}
\description{
find all the installed packages in the current system
}
