% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_version.R
\name{find_cran_version}
\alias{find_cran_version}
\title{find different versions available for a package in CRAN}
\usage{
find_cran_version(package = "data.table")
}
\arguments{
\item{package}{character\cr
package name}
}
\description{
find different versions available for a package in CRAN
}
