% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_version.R
\name{find_installed_location}
\alias{find_installed_location}
\title{find the location where package installed in current system}
\usage{
find_installed_location(package = "data.table")
}
\arguments{
\item{package}{character\cr
package name}
}
\description{
find the location where package installed in current system
}
