% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_version.R
\name{find_installed_package_version}
\alias{find_installed_package_version}
\title{find version of a package that is already installed in system}
\usage{
find_installed_package_version(package = "data.table")
}
\arguments{
\item{package}{character\cr
package name}
}
\description{
find version of a package that is already installed in system
}
