% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_functions.R
\name{finite_deriv}
\alias{finite_deriv}
\title{Find the numerical derivative of a finite set of points.}
\usage{
finite_deriv(x, y)
}
\arguments{
\item{x}{The x values for your set of points.}

\item{y}{The y vales for your set of points.}
}
\value{
A vector with the same length as x containing estimated derivative at
  each x value.
}
\description{
Uses the central difference method to calculate the numerical derivative for
a set of x,y data. Vectors x and y must be the same length.
}
