% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_spec.R
\name{fitIonInit}
\alias{fitIonInit}
\title{Initialize nls Fitting Using Ions}
\usage{
fitIonInit(
  ions,
  sd = 0.5,
  rel = NULL,
  peak = "gaussian",
  mass.shift = 0,
  noise = 0,
  tau = 0.1,
  charge = NULL,
  threshold = 10
)
}
\arguments{
\item{ions}{Character. A character vector of molecular formulae sutable to be
passed to \code{\link[enviPat]{isopattern}}.}

\item{sd}{Numeric. The estimated standard deviation (or equivalent for
peak = "emg") of the fitted peak.}

\item{rel}{Numeric. A numeric vector of relative heights for each ion to
initialize the starting values of the peak magnitudes.}

\item{peak}{Character. A string specifying what sort of peak shape to fit.
Acceptable values are "gaussian" (the default) and "emg". See Details.}

\item{mass.shift}{Numeric. The estimated mass shift of the measured mass
spectrum relative to the absolute mass position of the ions.}

\item{noise}{Numeric. The estimated value of the constant noise floor of the
mass spectrum.}

\item{tau}{Numeric. The estimated value of the skewness parameter for
peak = "emg". Ignored for peak = "gaussian".}

\item{charge}{Numeric.}

\item{threshold}{Numeric.}
}
\value{
A list containing elements of the fitting formula, the starting fit
values, and the lower limits for each parameter.
}
\description{
\code{fitIonInit} creates the formula and starting values for fitting mass
spectra using \code{\link[stats]{nls}} based on molecular formulae. The
isotopic signatures are generated using \code{\link[enviPat]{isopattern}}.
}
\details{
Fitting peak shape is either a gaussian (peak = "gaussian") or
exponentially modified gaussian (peak = "emg").
}
\seealso{
\code{\link[stats]{nls}}, \code{\link[enviPat]{isopattern}}

Other spectral fitting functions: 
\code{\link{ionFormula}()}
}
\concept{spectral fitting functions}
