% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{fragments}
\alias{fragments}
\title{Calculate all (combinatorically) possible fragments of a chemical formula.}
\usage{
fragments(chemform)
}
\description{
Calculate all (combinatorically) possible fragments of a chemical formula.
}
