% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{func.RCP.simple}
\alias{func.RCP.simple}
\title{Create  RCP Pattern with same intensity and number of points as experimental pattern}
\usage{
func.RCP.simple(
  exp.ppp,
  thickness,
  rcp.pattern,
  intensity.rcp = 1,
  maxGr,
  maxKr,
  nGr,
  nKr
)
}
\description{
Create  RCP Pattern with same intensity and number of points as experimental pattern
}
