% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{func_rcp_simple}
\alias{func_rcp_simple}
\title{Create  RCP Pattern with same intensity and number of points as experimental pattern}
\usage{
func_rcp_simple(
  exp_ppp,
  thickness,
  rcp_pattern,
  neighbor_number = 1,
  intensity_rcp = 1,
  maxGr,
  maxKr,
  nGr,
  nKr
)
}
\description{
Create  RCP Pattern with same intensity and number of points as experimental pattern
}
