% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_depracated.R
\name{func_rcp_simple_v01}
\alias{func_rcp_simple_v01}
\title{Create  RCP Pattern with same intensity and number of points as experimental pattern
DEPRACATED now just mulimersim() funciton with group_size = 1}
\usage{
func_rcp_simple_v01(
  exp_ppp,
  thickness,
  rcp_pattern,
  neighbor_number = 1,
  intensity_rcp = 1,
  maxGr,
  maxKr,
  nGr,
  nKr
)
}
\description{
Create  RCP Pattern with same intensity and number of points as experimental pattern
DEPRACATED now just mulimersim() funciton with group_size = 1
}
