% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{G3est_nn}
\alias{G3est_nn}
\title{Extended Nearest Neighbour Distance Function}
\usage{
G3est_nn(
  X,
  k = 1,
  rmax = NULL,
  nrval = 128,
  disjoint = NULL,
  correction = c("rs", "km", "han")
)
}
\arguments{
\item{X}{The observed point pattern, from which an estimate of the
kth nearest neightbor distance distribution function \eqn{G[3g](r)} will be computed.}

\item{k}{calculate distance to kth nearest neighbor}

\item{rmax}{Optional. Maximum value of argument \emph{r} for which \eqn{G[3g](r)}
will be estimated.}

\item{nrval}{Optional. Number of values of \emph{r} for which \eqn{G[3g](r)} will
be estimated. A large value of \code{nrval} is required to avoid discretisation
effects.}

\item{correction}{Optional. Character string specifying the edge
correction(s) to be used. Options are \code{"none"}, \code{"rs"}, \code{"km"},
\code{"hanisch"}, and \code{"best"}. Alternatively \code{correction="all"} selects all
options.}
}
\description{
For a marked point pattern, estimate the distribution of the distance from a
typical point to the kth nearest point of subset
}
\details{
The function \code{G3est_nn} extends \code{\link[spatstat.explore]{G3est}} to higher order NN's
}
\seealso{
\code{\link{G3multi}}, \code{\link{G3cross}}, \code{\link[spatstat.explore]{G3est}}

Other spatstat extensions: 
\code{\link{G3cross}()},
\code{\link{G3multi}()},
\code{\link{K3scaled}()},
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
