% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{g3features}
\alias{g3features}
\title{G features}
\usage{
g3features(rvals, gvals_new, gvals_old)
}
\arguments{
\item{rvals}{vector of radius values}

\item{gvals_new}{vector of new G3est values from which to subtract expected values}

\item{gvals_old}{vector of expected G3est values}
}
\value{
features to be used for machine learning prediction of clustering behavior.  The features are
the max difference between the functions (\emph{max_diff}), the r value at that difference (\emph{max_diff_r}),
the minimum difference (\emph{min_diff}), and the r value when the difference is 0 (\emph{zero_diff_r})
}
\description{
Features calculated from G function (G3est).
}
\details{
Takes one set of radius values and corresponding G values of two functions as input and returns
features for machine learning
}
