% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{G3multi}
\alias{G3multi}
\title{Marked Nearest Neighbour Distance Function}
\usage{
G3multi(
  X,
  I,
  J,
  rmax = NULL,
  nrval = 128,
  disjoint = NULL,
  correction = c("rs", "km", "han")
)
}
\arguments{
\item{X}{The observed point pattern, from which an estimate of the multitype
distance distribution function G3IJ(r) will be computed. It must be a
marked point pattern. See under Details.}

\item{I}{Subset of points of X from which distances are measured.}

\item{J}{Subset of points in X to which distances are measured.}

\item{rmax}{Optional. Maximum value of argument r for which G3IJ(r) will be
estimated.}

\item{nrval}{Optional. Number of values of r for which G3(r) will be
estimated. A large value of nrval is required to avoid discretisation
effects.}

\item{disjoint}{Optional flag indicating whether the subsets I and J are
disjoint. If missing, this value will be computed by inspecting the vectors
I and J.}

\item{correction}{Optional. Character string specifying the edge
correction(s) to be used. Options are "none", "rs", "km", "hanisch"
and "best". Alternatively correction="all" selects all options.}
}
\description{
For a marked point pattern, estimate the distribution of the distance from a
typical point in subset I to the nearest point of subset J.
}
