% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{g3Xfeatures}
\alias{g3Xfeatures}
\title{GX features}
\usage{
g3Xfeatures(rvals, gvals_new, gvals_old)
}
\arguments{
\item{rvals}{vector of radius values}

\item{gvals_new}{vector of new G3cross values from which to subtract expected values}

\item{gvals_old}{vector of expected G3cross values}
}
\value{
features to be used for machine learning prediction of clustering behavior.  The features are
the minimum difference between the functions (\emph{min_diff}),
the radius at which gvals_new = 0.95 \emph{ninty_fifth_percentile},
and the full width at half max of the peak formed by the difference in the functions \emph{FWHM}
}
\description{
Features calculated from G function (G3est).
}
\details{
Takes one set of radius values and corresponding G values of two functions as input and returns
features for machine learning
}
