% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-detection.R
\name{gema.gema}
\alias{gema.gema}
\title{Reprocess Clusters Identified Using GEMA}
\usage{
\method{gema}{gema}(X, threshold = 0.5)
}
\arguments{
\item{X}{An object of class "gema" (as produced by \code{\link{gema.pp3}})}

\item{threshold}{The probability threshold at which to assign a point to a
cluster or to the background.}
}
\value{
An
}
\description{
Reprocess Clusters Identified Using GEMA
}
\seealso{
Other cluster identification functions: 
\code{\link{gema.pp3}()},
\code{\link{gema}()},
\code{\link{msa}()}
}
\concept{cluster identification functions}
