% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-detection.R
\name{gema.pp3}
\alias{gema.pp3}
\title{Identify Clusters in a Point Pattern Using GEMA}
\usage{
\method{gema}{pp3}(X, cluster = NULL, kde.n = 20, max.clusters = 30, threshold = 0.5)
}
\arguments{
\item{X}{The point pattern (object of class \code{\link[spatstat]{ppp}} or
\code{\link[spatstat]{pp3}}) in which to identify clusters.}

\item{cluster}{The marks of \code{X} that are cluster-type points. If
\code{cluster = NULL} (the default) or \code{X} is unmarked, all points are assumed
to be cluster-type points.}

\item{kde.n}{Number of sampling grid lines for the 3D kernel density estimate}

\item{max.clusters}{The maximum number of clusters to define the cluster
search initialization. See Details.}

\item{threshold}{The probability threshold at which to assign a point to a
cluster or to the background. Defaults to 0.5, which is usually Bayes
optimal.}
}
\value{
A marked point pattern (of the same class as \code{X}) with marks
corresponding to the identified background and each identified cluster
based on \code{threshold}.
}
\description{
Identify Clusters in a Point Pattern Using GEMA
}
\details{
A \code{data.frame} is returned as an attribute (\code{attr("prob")}) that contains the
probabilities of assignment to the background and each cluster.
}
\references{
Zelenty, J. \emph{et al.},
"Detecting Clusters in Atom Probe Data with Gaussian Mixture Models",
\emph{Microscopy and Microanalysis}, \strong{23} (2), 269-278 (2017):
\url{https://doi.org/10.1017/S1431927617000320}
}
\seealso{
Other cluster identification functions: 
\code{\link{gema.gema}()},
\code{\link{gema}()},
\code{\link{msa}()}
}
\concept{cluster identification functions}
