% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-detection.R
\name{gema}
\alias{gema}
\title{Identify Clusters in a Point Pattern Using GEMA}
\usage{
gema(X, ...)
}
\description{
Identify Clusters in a Point Pattern Using GEMA
}
\seealso{
Other cluster identification functions: 
\code{\link{gema.gema}()},
\code{\link{gema.pp3}()},
\code{\link{msa}()}
}
\concept{cluster identification functions}
