% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_multi.R
\name{indexMultiples}
\alias{indexMultiples}
\title{Returns a list of indicies where the position in the list is the hit
multiplicity and the index is of the first hit in the multiple hit event.}
\usage{
indexMultiples(ato, cl.n = 3)
}
\arguments{
\item{ato}{An ATO data frame. The data frame should be generated by
\code{\link{readATO}}.}

\item{cl.n}{An integer. The number of cores to use when creating the parallel
cluster with \code{\link[parallel]{makeCluster}}.}
}
\description{
\code{indexMultiples} extracts the indicies of the \emph{first} hit for every
multiplicity in the ATO file. The indicies are sorted into a list that is
ordered by multiplicity. This list is then returned.
}
\details{
This function uses the \code{parallel} package to speed processing of the ATO
file. The argument \code{cl.n} defines the number of cores to use according
to \code{\link[parallel]{makeCluster}}.
}
\seealso{
\code{\link{multiples}}, \code{\link{readATO}},
  \code{\link[parallel]{makeCluster}}.
}
