% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{inside.pp3}
\alias{inside.pp3}
\title{Extends \code{\link[spatstat]{inside}} to \code{\link[spatstat]{pp3}}.}
\usage{
inside.pp3(points, domain = NULL)
}
\description{
Extends \code{\link[spatstat]{inside}} to \code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[spatstat]{inside}}
}
