% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_version.R
\name{install_package_version}
\alias{install_package_version}
\title{interactively select which version to install for a package}
\usage{
install_package_version(package)
}
\arguments{
\item{package}{character\cr}
}
\description{
find R version history
}
