% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{intensity.pp3-deprecated}
\alias{intensity.pp3-deprecated}
\title{Extends \code{\link[spatstat.geom]{intensity}} to \code{\link[spatstat.geom]{pp3}}.}
\description{
Extends \code{\link[spatstat.geom]{intensity}} to \code{\link[spatstat.geom]{pp3}}.
}
\seealso{
\code{\link{rapt-deprecated}}
}
\keyword{internal}
