% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spat.R
\name{intensity.pp3}
\alias{intensity.pp3}
\title{Extends \code{\link[spatstat]{intensity}} to \code{\link[spatstat]{pp3}}.}
\usage{
intensity.pp3(X, weights = NULL)
}
\description{
Extends \code{\link[spatstat]{intensity}} to \code{\link[spatstat]{pp3}}.
}
