% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{isoLabel}
\alias{isoLabel}
\title{Label a prettyPlot with isotope patterns generated by
\code{\link[enviPat]{isopattern}}}
\usage{
isoLabel(ms, pat, rng = NULL, col = NULL, halfWindowSize = 0.5)
}
\arguments{
\item{ms}{A MassSpectrum. The (log transformed) mass spectrum to be marked.}

\item{pat}{A list of isotopic patterns as generated by
\code{\link[enviPat]{isopattern}}}

\item{rng}{Currently unused.}

\item{col}{A vector of colors for the isotopes. If NULL, the standard R
sequence of collors will be used.}

\item{halfWindowSize}{A numeric. The half-width of the window for determining
the height of the main isotope peak.}
}
\description{
Label a prettyPlot with isotope patterns generated by
\code{\link[enviPat]{isopattern}}
}
\seealso{
\code{\link[enviPat]{isopattern}}
}
