% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{k3features}
\alias{k3features}
\title{Extract features from 3D K function output
This is the same as k3metrics.  This function has update terminology ("metric" to "feature") to be more accurate}
\usage{
k3features(rvals_new, tvals_new, toplot, ...)
}
\arguments{
\item{rvals_new}{The radius values from the \code{\link{anomK3est}}. These
need to be trimmed so that they start past the inhibition radius of the
underlying RCP pattern.}

\item{tvals_new}{The returned K values from the \code{\link{anomK3est}}.
These need to be trimmed to the same indices as rvals_new.}
}
\value{
A list of [\link{1}] Km, [\link{2}] Rm, [\link{3}] Rdm, [\link{4}] Rddm, [\link{5}] Kdm for
the input K function. If no first K peak is found, returns a list of 5
NaNs.
}
\description{
Takes as inputs the results from \code{\link{anomK3est}} function performed
on a clustered data set. Returns five different features of this result.
}
\seealso{
\code{\link{anomK3est}}
}
