% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_K3metrics.R
\name{k3metrics}
\alias{k3metrics}
\title{Extract metrics from 3D K function output}
\usage{
k3metrics(rvals.new, tvals.new, toplot, ...)
}
\arguments{
\item{rvals.new}{The radius values from the \code{\link{anomK3est}}. These
need to be trimmed so that they start past the inhibition radius of the
underlying RCP pattern.}

\item{tvals.new}{The returned K values from the \code{\link{anomK3est}}.
These need to be trimmed to the same indices as rvals.new.}
}
\value{
A list of [[1]] Km, [[2]] Rm, [[3]] Rdm, [[4]] Rddm, [[5]] Kdm for
  the input K function. If no first K peak is found, returns a list of 5
  NaNs.
}
\description{
Takes as inputs the results from \code{\link{anomK3est}} function performed
on a clustered data set. Returns five different metrics of this result.
}
\seealso{
\code{\link{anomK3est}}
}
