% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{K3scaled}
\alias{K3scaled}
\title{Locally Scaled K-function in Three Dimensions}
\usage{
K3scaled(
  X,
  lambda = NULL,
  ...,
  rmax = NULL,
  nrval = 128,
  correction = "none",
  renormalise = FALSE,
  normpower = 1
)
}
\arguments{
\item{X}{A pp3}

\item{lambda}{An estimate of the intensity function f(x,y,z). Defaults to a
uniform intensity, estimated with \code{\link[spatstat]{intensity}}; this
results in the same behavior as \code{\link[spatstat]{K3est}}.}

\item{correction}{The correction to be used. Currently only the uncorrected
estimate (\code{correction = "none"}) is implemented.

In 3D the cube root is used instead of the square root to estimate the
correction of the scaled intensity. Thas is, the eucledian distance between
two points \eqn{u[i]} and \eqn{u[j]}, \eqn{d[ij]}, is
multiplied by the mean of the cube roots of the local intensities
\eqn{lambda[c](u[i])} and \eqn{lambda[c](u[j])}.}
}
\description{
Estimates the locally-rescaled K-function of a 3D point pattern
}
\seealso{
\code{\link[spatstat]{Kscaled}}

Other spatstat extensions: 
\code{\link{G3cross}()},
\code{\link{G3multi}()},
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
