% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_K3metrics.R
\name{kseries}
\alias{kseries}
\title{Simulate a clustered point pattern and extract the K-function metrics from
it.}
\usage{
kseries(j, params, maxr, nr, toSub,
  rcp_path = "~/Research/point_patterns/Final", s = NULL)
}
\arguments{
\item{j}{The index of the RCP pattern to run.}

\item{params}{Vector of length 3 or 4 containing either \code{c(r, den, gbp)}
or \code{c(r, den, rb, gbp)}.}

\item{maxr}{Maximum r to calculate K-function to.}

\item{nr}{Number of r values to evaluate K-function at.}

\item{toSub}{A vector of values to substitute for the \code{\link{anomK3est}}
function.}

\item{rcp_path}{String holding the file path to the directory holding the RCP
'FinalConfig' and 'system' files.}

\item{s}{Random seed.}
}
\value{
Vector containing 5 metrics.
}
\description{
This function takes an RCP point pattern and a single set of parameters to
create a cluster and extracts the K function metrics.
}
