% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_K3metrics.R
\name{kseries2}
\alias{kseries2}
\title{Helper function to allow for parallelization of metric extraction from
cluster simulations}
\usage{
kseries2(j, p, tot, maxr, nr, toSub,
  rcp_path = "~/Research/point_patterns/Final", verbose = FALSE,
  junk_path = "~/Research/junk/", s)
}
\arguments{
\item{j}{The index of the RCP pattern to run (note that this might only be
able to be applied on the mines HPC under Galen's account right now as
written)}

\item{p}{The total number of RCP patterns available.}

\item{tot}{A list of cluster properties to test. This should be a list of
vectors containing c(r, den, rb, gbp).}

\item{maxr}{Maximum r to calculate K function to.}

\item{nr}{Number of r values to evaluate k function at.}

\item{toSub}{A vector of values to substitute for the \code{\link{anomK3est}}
function.}

\item{rcp_path}{String holding the file path to the directory holding the RCP
'FinalConfig' and 'system' files.}

\item{verbose}{\code{TRUE} or \code{FALSE}. Whether to output update files to
a junk folder.}

\item{junk_path}{String holding the file path to the directory where the
update files should go if \code{verbose = TRUE}.}

\item{s}{Random seed.}
}
\value{
Matrix containing 5 metrics and the seed value for each parameter
  combination given in the \code{tot} list.
}
\description{
This function takes a single RCP pattern and creates and tests the K function
metrics on different cluster properties passed into the function. Used to
parallelize large sets of cluster property runs.
}
