% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_K3metrics.R
\name{kseries3}
\alias{kseries3}
\title{Helper function to allow for parallelization of metric extraction from
cluster simulations (the newest ones)}
\usage{
kseries3(j, p, tot, maxr, nr, toSub, pcp = 0.1, tol = 0.005,
  rcp_path = "~/Research/point_patterns/Final", verbose = FALSE,
  junk_path = "~/Research/junk/", s = NULL)
}
\arguments{
\item{j}{The index of the RCP pattern to run.}

\item{p}{The total number of RCP patterns available.}

\item{tot}{A list of cluster properties to test. This should be a list of
vectors containing \code{c(r, den, rb, gbp)} OR \code{c(r, den, gbp)}.}

\item{maxr}{Maximum r to calculate K function to.}

\item{nr}{Number of r values to evaluate k function at.}

\item{toSub}{A vector of values to substitute for the \code{\link{anomK3est}}
function.}

\item{pcp}{See \code{pcp} argument for \code{\link{makecluster}} function.}

\item{rcp_path}{String holding the file path to the directory holding the RCP
'FinalConfig' and 'system' files.}

\item{verbose}{\code{TRUE} or \code{FALSE}. Whether to output update files to
a junk folder.}

\item{junk_path}{String holding the file path to the directory where the
update files should go if \code{verbose = TRUE}.}

\item{s}{Random seed.}
}
\value{
Matrix containing 5 metrics for each parameter combination given in
  the \code{tot} list.
}
\description{
This function takes a single RCP pattern and creates and tests the K function
metrics on different cluster properties passed into the function. Used to
parallelize large sets of cluster property runs.
}
