% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim.R
\name{lattice}
\alias{lattice}
\title{Creates a spatial region filled with lattice points of the specified type}
\usage{
lattice(domain = box3(), a = 1, lattice = "sc")
}
\description{
Creates a spatial region filled with lattice points of the specified type
}
