% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim.R
\name{latticeVectors}
\alias{latticeVectors}
\title{Generate Lattice Vectors}
\usage{
latticeVectors(indices, a = 1, lattice = "sc")
}
\description{
Transform a matrix of integer indices to spatial coordinates for a specified
lattice type
}
