% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_spat.R
\name{lineProfile.pp3}
\alias{lineProfile.pp3}
\title{Calculate a 1D Concentration or Density Profile}
\usage{
\method{lineProfile}{pp3}(
  X,
  mark = NULL,
  concentration = TRUE,
  axis = "z",
  np = 10,
  spacing = "width",
  breaks = NULL,
  simplify = TRUE
)
}
\arguments{
\item{X}{The \code{\link[spatstat]{pp3}} object to split up.}

\item{concentration}{Should the relative concentration of the marks be
returned?}

\item{axis}{The axis of analysis.}

\item{np}{The number of sample points along the analysis axis. Incompatible
with \code{breaks}.}

\item{spacing}{The spacing used for the concentration sampling: one of
\code{"width"} (the default) or \code{"number"}. \code{"width"} uses sampling bins that
are equal divisions of the analysis axis, while \code{"number"} uses bins that
contain equal numbers of points. Incompatible with \code{breaks}.}

\item{breaks}{The breakpoints of the sampling bins. If \code{NULL} (the default),
the breaks will be calculated based on \code{np} and \code{spacing}.}
}
\description{
Along a specified axis, divide a 3D point pattern into samples to calculate
either the absolute point density or the concentration relative to a
specified reference.
}
