% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_NNDE.R
\name{local.den.engine}
\alias{local.den.engine}
\title{Find the local intensity density estimate based on nearest neighbors.}
\usage{
local.den.engine(bdist, nnk, k, dz, par = TRUE, cores = 7)
}
\arguments{
\item{bdist}{Result from \code{\link{bdist.points3.multi}} giving shortest
distance to boundary in the x, y, and z directions.}

\item{nnk}{Result from \code{\link[spatstat]{nndist}} or
\code{\link[spatstat]{nncross}} containing nearest neighbor distances.}

\item{k}{Vector of nn#s to calculate estimate for.}

\item{dz}{The spacing for numeric integration over the z direction to
determine edge corrections.}

\item{par}{\code{TRUE} or \code{FALSE}: whether or not to calculate in
parallel}

\item{cores}{If \code{par = TRUE}, this is the number of cores to use for the
parallel calculation.}
}
\value{
A data.frame with intensity estimates for each value of k (# of
  nearest neighbors).
}
\description{
Helper for \code{\link{nndensity.pp3}}.
}
\seealso{
\code{\link{nndensity.pp3}}
}
