% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_localK3.R
\name{localk3engine}
\alias{localk3engine}
\title{Wrapper function for the C code under \code{\link{localK3est}}.}
\usage{
localk3engine(x, y, z, box = c(0, 1, 0, 1, 0, 1), rmax = 1, nrval = 100)
}
\description{
Wrapper function for the C code under \code{\link{localK3est}}.
}
