% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_K3metrics.R
\name{localk3metrics}
\alias{localk3metrics}
\title{Extract variance of R_max over local K functions from a point pattern.}
\usage{
localk3metrics(kl, start, nsamp = NULL)
}
\arguments{
\item{kl}{The output object from a \code{\link{localK3est}} function call.}

\item{start}{The x value index to start searching for mertics at (to avoid
strange behavior near r = 0).}

\item{nsamp}{How many samples of R_max to take from the sample. If
\code{NULL}, defaults to the number of points in the pattern (this can be
very computationally expensive for large patterns).}
}
\value{
A signle numerical value which is the variance of the R_max metric
  over \code{nsamp} individual points in the pattern.
}
\description{
This function takes the output of \code{\link{localK3est}}, calculates R_max
for the individual local K measures around each point, and returns the
variance of these measures.
}
\seealso{
\code{\link{localK3est}}, \code{\link[spatstat]{K3est}},
  \code{\link{kseries2}}
}
