% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim-cluster.R
\name{makeCluster}
\alias{makeCluster}
\title{Simulate Clusters}
\usage{
makeCluster(under, over, radius1, radius2, type = "ppc", ppc = NULL,
  cr = NULL, d = NULL, pic = 1, pcp = 0.06, toPlot = FALSE,
  showOverPts = FALSE)
}
\arguments{
\item{under}{A pp3. The underlying point pattern}

\item{over}{A pp3. The pattern of clusters}

\item{radius1}{A numeric. The small radius of the rcpUnder pattern}

\item{radius2}{A numeric. The small radius of the rcpOver pattern}

\item{type}{A string. One of "ppc" (points per cluster), "cr" (cluster
radius), or "dist" (distance between points). See Details.}

\item{ppc}{A numeric. The points per cluster that you want for "ppc." This is
a SUGGESTION; it will get close.}

\item{cr}{A numeric. The radius of cluster that you want for "cr."}

\item{d}{A numeric. The distance between clusters that you want for "dist."}

\item{pic}{A numeric. The % of the points designated as cluster type points
that you would like to actually be in clusters.}

\item{pcp}{A numeric. The % of points out of the underlying pattern that you
want to be included as points of the clustering type.}
}
\description{
\code{makeCluster} uses two point patterns to generate clusters. One is the
underlying point pattern, from which the actual simulated data will be taken.
The other is to randomly place the cluster locations through the underlying
pattern.
}
