% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-extend.R
\name{marktable.pp3}
\alias{marktable.pp3}
\title{Extends \code{\link[spatstat]{marktable}} to \code{\link[spatstat]{pp3}}.}
\usage{
marktable.pp3(X, R, N, exclude = TRUE, collapse = FALSE)
}
\description{
\code{marktable.pp3}
}
