% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{marktable.ppp}
\alias{marktable.ppp}
\title{Tabulate Marks in Neighbourhood of Every Point in a Point Pattern}
\usage{
\method{marktable}{ppp}(X, R, N, exclude = TRUE, collapse = FALSE)
}
\description{
Tabulate Marks in Neighbourhood of Every Point in a Point Pattern
}
\seealso{
\code{\link[spatstat]{marktable}}
}
