% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{marktable}
\alias{marktable}
\title{Tabulate Marks in Neighbourhood of Every Point in a Point Pattern}
\usage{
marktable(X, ...)
}
\description{
This is an S3 generic that extends the use of
  \code{\link[spatstat]{marktable}} beyond \code{ppp} objects.
}
\seealso{
\code{\link[spatstat]{marktable}}, \code{\link{marktable.ppp}},
  \code{\link{marktable.pp3}}

Other spatstat extensions: 
\code{\link{Tstat.pp3}()},
\code{\link{marktable.pp3}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
