% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{morph_gb}
\alias{morph_gb}
\title{Simulate morphology of guest molecules on a grain boundary}
\usage{
morph_gb(
  lambda,
  frac,
  point.den = 1,
  rcp.rad,
  rcp.path = NULL,
  rcp.number = 1,
  toplot = FALSE,
  win = box3(c(0, 1), c(0, 1), c(0, 1))
)
}
\arguments{
\item{lambda}{Intensity for the background Poisson process.}

\item{frac}{Fraction of points to select as guest points. Between zero and
one.}

\item{point.den}{Point density within the volume of the guest volume.}

\item{rcp.rad}{Radius to scale the RCP pattern to.}

\item{rcp.path}{File path to folder containing the 3D RCP patterns which
contain both the 'FinalConfigx' and 'systemx' files.}

\item{rcp.number}{Which RCP file to use.}

\item{toplot}{\code{TRUE} or \code{FALSE}; Plot results or not.}

\item{win}{The simulation window.}
}
\value{
A list of: [[1]] a \code{\link[spatstat]{pp3}} object contining the
  guest points. [[2]] A \code{pp3} object containing the entire background
  underlying point pattern.
}
\description{
Simulate morphology of guest molecules on a grain boundary
}
