% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{morph_gyroid}
\alias{morph_gyroid}
\title{Simulate morphology of guest molecules in gyroid pattern}
\usage{
morph_gyroid(
  lambda,
  frac,
  gyroid.scale,
  point.den = 1,
  toplot = FALSE,
  win = box3(c(0, 1), c(0, 1), c(0, 1))
)
}
\arguments{
\item{lambda}{Intensity for the background Poisson process.}

\item{frac}{Fraction of points to select as guest points. Between zero and
one.}

\item{gyroid.scale}{Scale factor for the gyroid pattern. Larger = tighter
spacing of the gyroid.}

\item{point.den}{Point density within the volume of the guest volume.}

\item{toplot}{\code{TRUE} or \code{FALSE}; Plot results or not.}

\item{win}{The simulation window.}
}
\value{
A list of: [[1]] a \code{\link[spatstat]{pp3}} object contining the
  guest points. [[2]] A \code{pp3} object containing the entire background
  underlying point pattern.
}
\description{
Simulate morphology of guest molecules in gyroid pattern
}
