% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{morph_lamellar}
\alias{morph_lamellar}
\title{Simulate lamellar morphology for mixed guest-host system.}
\usage{
morph_lamellar(
  lambda,
  frac,
  plane.norm = c(1, 0, 0),
  plane.den,
  point.den = 1,
  toplot = FALSE,
  win = box3(c(0, 1), c(0, 1), c(0, 1))
)
}
\arguments{
\item{lambda}{Intensity for the background Poisson process.}

\item{frac}{Fraction of points to select as guest points. Between zero and
one.}

\item{plane.norm}{Vector you wish to define normal of the planes with.
Doesn't need to be normalized. Positive values only}

\item{plane.den}{Linear plane density.}

\item{point.den}{Point density within the plane volumes.}

\item{toplot}{\code{TRUE} or \code{FALSE}. Plot results or not.}

\item{win}{The simulation window.}
}
\value{
A list of: [\link{1}] a \code{\link[spatstat]{pp3}} object contining the
guest points. [\link{2}] A \code{pp3} object containing the entire background
underlying point pattern.
}
\description{
Simulate lamellar morphology for mixed guest-host system.
}
\seealso{
Other simulation functions: 
\code{\link{clustersim}()},
\code{\link{lattice}()},
\code{\link{makecluster}()},
\code{\link{morph_rods}()},
\code{\link{nmers}()},
\code{\link{read.rcp}()},
\code{\link{scaleRCP}()},
\code{\link{stitch.size}()}

Other morphology generators: 
\code{\link{morph_rods}()}
}
\concept{morphology generators}
\concept{simulation functions}
