% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{morph_rods}
\alias{morph_rods}
\title{Simulate morphology of guest rods in a two-phase system}
\usage{
morph_rods(
  lambda,
  frac,
  rod.norm = "z",
  rod.den,
  rod.spacing = "grid",
  rcp.path = NULL,
  point.den = 1,
  toplot = FALSE,
  win = box3(c(0, 1), c(0, 1), c(0, 1))
)
}
\arguments{
\item{lambda}{Intensity for the background Poisson process.}

\item{frac}{Fraction of points to select as guest points. Between zero and
one.}

\item{rod.norm}{Either 'x', 'y', or 'z'; defines which axis the rods should
be alligned with.}

\item{rod.den}{Rod density; 2D vector containing a linear density for both
directions.}

\item{rod.spacing}{One of: "grid", "rcp", "hexagonal", or "random". The
method used to space the rods.}

\item{rcp.path}{If \code{rod.spacing} is set to 'rcp', then set the file path
to the folder contining the rcp FinalConfig and system files}

\item{point.den}{Point density within the volume of the rods.}

\item{toplot}{\code{TRUE} or \code{FALSE}; Plot results or not.}

\item{win}{The simulation window.}
}
\value{
A list of: [[1]] a \code{\link[spatstat]{pp3}} object contining the
  guest points. [[2]] A \code{pp3} object containing the entire background
  underlying point pattern.
}
\description{
Simulate morphology of guest rods in a two-phase system
}
\seealso{
Other simulation functions: 
\code{\link{clustersim}()},
\code{\link{lattice}()},
\code{\link{makecluster}()},
\code{\link{morph_lamellar}()},
\code{\link{nmers}()},
\code{\link{read.rcp}()},
\code{\link{scaleRCP}()},
\code{\link{stitch.size}()}

Other morphology generators: 
\code{\link{morph_lamellar}()}
}
\concept{morphology generators}
\concept{simulation functions}
