% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-detection.R
\name{msa}
\alias{msa}
\title{Identify Clusters in a Marked Point Pattern Using MSA}
\usage{
msa(X, dmax, Nmin, denv, der, clust.mark = c("A"))
}
\arguments{
\item{X}{A marked \code{\link[spatstat]{pp3}} object on which MSA will be
performed.}

\item{dmax}{The maximum distance two points can be separated by and still be
considered part of the same cluster.}

\item{Nmin}{The minimum number of points needed to classify a grouping of
points as a cluster.}

\item{denv}{Any points within this distance of a point residing in a cluster
will be included in the cluster (this controls the addition of background
points to a cluster).}

\item{der}{Any points within this distance of a background matrix point
(after enveloping) will be removed from the cluster.}

\item{clust.mark}{Vector containing the names of the marks in \code{X} that
should be considered as cluster type points. All points with marks not
included in this vector will be considered background points.}
}
\value{
A list of:
\itemize{
\item \code{radius} - A vector containing estimated radius of each cluster found
\item \code{den} - A vector containing estimated intra-cluster concentration of cluster
type points in each cluster found
\item \code{bgnd.den} - The estimated background concentration of cluster type points
within the background (\emph{i.e.} the entire domain minus the clusters)
\item \code{cluster} - A list of indices from the original pattern of cluster type
points that reside in clusters
\item \code{bgnd} - A list of indices from the original pattern of background type
points that reside in clusters.
}
}
\description{
\code{msa} segments a marked \code{\link[spatstat]{pp3}} into clusters and
background matrix using the maximum separation algorithm (MSA. The marks can
have more than two types, but MSA requires that each type is categorized as
either a cluster or non-cluster species.
}
\references{
Marquis, E.A. & Hyde, J.M.,
"Applications of atom-probe tomography to the characterisation of solute
behaviours,"
\emph{Materials Science and Engineering: R: Reports}, \strong{69} (4-5), 37-62 (2010):
\url{https://doi.org/10.1016/j.mser.2010.05.001}
}
