% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{multimers_relab}
\alias{multimers_relab}
\title{Relabel multimers}
\usage{
multimers_relab(
  exp.ppp,
  num.relabs,
  thickness,
  rcp.list,
  maxGr,
  maxKr,
  nGr,
  nKr,
  group_size = 2,
  num_neighbors = 6,
  weights = c(1, 1, 1/4),
  probs = c(0.6, 0.2, 0.2, 0),
  ncores = detectCores()
)
}
\description{
Relabel multimers
}
