% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_02.R
\name{multimersim}
\alias{multimersim}
\title{Multimer Simulation}
\usage{
multimersim(
  dopant_pattern,
  upp,
  min_thick = 0,
  max_thick = 10,
  group_size = 2,
  num_neighbors = 6,
  weights = c(1, 1, 1),
  probs = c(1, 0, 0, 0),
  intensity_upp = NULL
)
}
\arguments{
\item{dopant_pattern}{point pattern of class \emph{ppp} or \emph{pp3}.  The final multimer
pattern will match this pattern in class, intensity, and domain.}

\item{upp}{point pattern of class \emph{ppp} or \emph{pp3} to use as underlying point pattern.
Multimers will be selected from these points.}

\item{min_thick}{if \emph{dopant_pattern} is 2d (ppp) and \emph{upp} is 3d (pp3) this
determines the smallest z value to keep before collapsing into 2d.}

\item{max_thick}{if \emph{dopant_pattern} is 2d (ppp) and \emph{upp} is 3d (pp3) this
determines the largest z value to keep before collapsing into 2d.}

\item{group_size}{size of clusters.  If equal to 1, then all points will be independent
of each other}

\item{num_neighbors}{number of nearest neighbors to select from when
forming dimers, trimers, etc..}

\item{weights}{vector of length equal to number of dimensions in \emph{upp}. the weighted distance to each of \emph{num_neighbors} nearest neighbors
is calculated using \eqn{\sqrt{w_1 x^2 + w_2 y^2 + w_3 z^2}},
 where \emph{weights} = (\eqn{w_1, w_2, w_3})}

\item{probs}{vector of probabilities.  Should sum to \emph{group_size}-1.
For \eqn{probs = c(p_1, p_2, p_3, p_4)}, the probability of the first NN being
selected in \eqn{p_1}, the probability of the second is \eqn{p_2}, and so on}

\item{intensity_upp}{the \emph{upp} will be rescaled to have this intensity before the
marks are assigned.}
}
\description{
Simulates multimers (groups of two/dimers, three/trimers, etc.) in
underlying point pattern (upp) to match domain and number of points in \emph{dopant_pattern}.
}
\details{
algorithm steps:
 \itemize{
 \item{"Step 1"}{rescale \emph{upp} to match \emph{intensity_upp}}
 \item{"Step 2"}{Select points in the scaled \emph{upp} that are
 inside the domain of \emph{dopant_pattern}. }
}
}
