% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_depracated.R
\name{multimersim_v01}
\alias{multimersim_v01}
\title{Simulate 2D Multimers}
\usage{
multimersim_v01(
  exp_ppp,
  thickness,
  group_size = 2,
  num_neighbors = 6,
  neighbor_number = 1,
  weights = c(1, 1, 1/4),
  probs = c(0.6, 0.2, 0.2, 0),
  rcp_pattern,
  intensity_rcp = 1,
  maxGr,
  maxKr,
  nGr,
  nKr
)
}
\description{
Simulate
}
