% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_02.R
\name{neighbors}
\alias{neighbors}
\title{Neighbors in shell}
\usage{
neighbors(
  pattern,
  type = NULL,
  window = NULL,
  k_vec = 1,
  drop_isolated = TRUE,
  ...
)
}
\arguments{
\item{pattern}{point pattern of class \emph{ppp} or \emph{pp3}}

\item{type}{mark defining point type}

\item{window}{object of class \emph{owin} (for ppp) or \emph{box3} (for pp3).
Only points inside the window are used to calculate distances from, but points outside the
window are still included as potential neighbors. If \emph{NULL} then it includes the
entire domain of \emph{pattern}}

\item{k_vec}{an integer or vector of integers that determines which NN's to calculate}

\item{drop_isolated}{if \emph{TRUE} then points that are closer to a boundary than they
are to their \emph{kth} NN will be dropped: their distances are set to \emph{NA}}
}
\description{
Compare distribution of marks to binomial distribution
}
\details{
For each k in \emph{k_vec}, calculate how many of the points in
\emph{pattern} with mark \emph{type} have  0, 1, .. \emph{k} points of mark \emph{type} in their \emph{k}
nearest neighbors.
This can then be compared to the expected values of a binomial distribution with n = k and
p = fraction of points in \emph{pattern} that have mark \emph{type}
}
