% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim.R
\name{nmers}
\alias{nmers}
\title{Select n-mer Clusters}
\usage{
nmers(sam, parent, n = 2, full = FALSE)
}
\arguments{
\item{sam}{pp3. The seed points for the n-mers}

\item{parent}{pp3. The point pattern from which to draw the n-mers}

\item{n}{numeric. The number of points in the n-mer. Default is 2}

\item{full}{logical. Return a marked pattern with both n-mers and
"background" points? Default is \code{FALSE}}
}
\value{
A pp3 with the n-mers. If \code{full = FALSE} (the default) this is
an unmarked pattern with only the positions of the n-mers. If
\code{full = TRUE}, this is a marked pattern with marks "nmers" and "bkgd".
}
\description{
Creates an n-mer point pattern by selecting the n+1 nearest neighbors of a
sampled point pattern from its parent distribution.
}
\seealso{
\code{\link{clustersim}}

Other simulation functions: 
\code{\link{clustersim}()},
\code{\link{lattice}()},
\code{\link{makecluster}()},
\code{\link{morph_lamellar}()},
\code{\link{morph_rods}()},
\code{\link{read.rcp}()},
\code{\link{scaleRCP}()},
\code{\link{stitch.size}()}
}
\concept{simulation functions}
