% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_02.R
\name{nn_binom}
\alias{nn_binom}
\title{nn_binom}
\usage{
nn_binom(k_vec, p)
}
\arguments{
\item{k_vec}{vector of integers to use as \emph{n} values for binomial distribution}

\item{p}{probability of success. Should equal the concentration of the point pattern that
this will be compared to using the \emph{neighbors()} function}
}
\description{
calculates binomial distributions with \emph{n = k}
for each \emph{k} in \emph{k_vec} and probability of success \emph{p}
}
