% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndist_analysis_functions.R
\name{nn_dist_perc_table}
\alias{nn_dist_perc_table}
\title{Plot and Save Histograms of nndist function}
\usage{
nn_dist_perc_table(
  image,
  image2 = NULL,
  value = "percent",
  rads_vec = 1:25,
  nn_vec = 1:5,
  perc_vec = seq(0, 1, 0.1),
  output = "plot",
  location = getwd(),
  image_name = "image",
  unit = "nm",
  round_to = 2,
  ...
)
}
\value{
returns at matrix. If `value` = "percent", then rows are NN number, columns are distance, and value is percent.  If `value` = "distance",
 rows are NN number, columns are percent, and `value` is distance.  If `output` = "save" then it will also save the image at location `location`
}
\description{
calculate percent of each nearest neighbor (NN) in `nn_vec` that are less than x `unit` in `rads_vec` apart if `value` = "percent" or calculate the distance
at which each percentage in `perc_vec` has each NN in `nn_vec` if `value` = "distance."
}
\details{
@param image Image of type ppp or pp3 to calculate distances from
 @param image2 will be NULL (default) if using \code{\link[spatstat]{nndist}} function on `image`.  Otherwise, set as same class as `image` (ppp or pp3) to
 calculate distances to using (\code{\link[spatstat]{nncross}})
 @param value determines the output.  If = "percent", then rows are NN number, columns are distance, and value is percent.  If = "distance",
 rows are NN number, columns are percent, and `value` is distance
 @param rads_vec vector of distances to calculate percents at. Only used if `value` = "percent"
 @param nn_vec vector of integers values to determine which nearest neighbors to calculate
 @param perc_vec vector of percent values between 0 and 1 (inclusive) to calculate distances for.  Only used if `value` = "distance"
 @param output set to "plot" (default) if only plotting.  Set to "save" if plotting and saving
 @param location location to save image to if `output = save`.  default is working directory.
 @param image_name Name to save image as.  Will have `_NN_dist_perc.png` added to it if `value` = "percent" or `_NN_perc_dist.png` if `value` = "distance"
 @param unit distance unit name to be added to plot name
 @param round_to number of decimal places to round values to


 add details
}
