% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_NNDE.R
\name{nndensity.pp3}
\alias{nndensity.pp3}
\title{Extension of \code{\link[spatstat]{nndensity}} to handle pp3 objects.}
\usage{
\method{nndensity}{pp3}(X, k, nx, ny, nz, dz, at.points = FALSE, par = TRUE, cores = 7)
}
\arguments{
\item{X}{The point pattern to estimate the intensity of.}

\item{k}{Vector containing the nearest-neighbor #s that the estimate should
be calculated for.}

\item{nx, ny, nz}{If estimating on a grid, the number of grid points in x, y,
and z.}

\item{dz}{The z spacing for numeric integration for the edge correction
(suggested ~ 0.1-0.5)}

\item{at.points}{\code{TRUE} or \code{FALSE}. Whether or not to estimate
intensity at points in pattern. If \code{TRUE}, nx, ny, and nz are not
used.}

\item{par}{\code{TRUE} or \code{FALSE}: whether or not to calculate in
parallel}

\item{cores}{If \code{par = TRUE}, this is the number of cores to use for the
parallel calculation.}
}
\value{
List containing: [\link{1}] A data frame of the intensity estimates for
each nearest neighbor value. [\link{2}] The coordinates of the estimates. [\link{3}]
The coordinates of the original points from the data set.
}
\description{
Calculates the 3D nearest-neighbor intensity estimate of a point process at
either a grid of points or at the point locations in the data set. Utilizes
the volume weighted edge correction. See Statistics for Spatial Data by
Cressie pg. 654 for more info.
}
