% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nndist_analysis_functions.R
\name{nndist_func_plotter}
\alias{nndist_func_plotter}
\title{nndist plotter and saver}
\usage{
nndist_func_plotter(
  image,
  image2 = NULL,
  nn_vec = 1:5,
  location = getwd(),
  image_name = "image",
  output = "save",
  xlim = c(0, 10),
  ylim = c(0, 0.5),
  binSize = 0.5,
  freq = FALSE,
  image_width = 600,
  image_height = 500,
  ...
)
}
\arguments{
\item{image}{Point pattern for analysis (ppp or pp3)}

\item{image2}{If a point pattern is given for image2, then \code{\link[spatstat]{nncross}} will be used}

\item{nn_vec}{vector of nearest neighbors to calculate distance to}

\item{location}{location to save image(s) if `output` = `save`}

\item{image_name}{Name to save image as.  The nearest neighbor number + `nn.png` will be added to it}

\item{xlim}{limits of x axis of histogram}

\item{ylim}{limits of y axis of histogram}

\item{binSize}{histogram bin size}

\item{freq}{argument for  \code{\link{hist}} function. logical;
if TRUE, the histogram graphic is a representation of frequencies, the counts component of the result;
if FALSE, probability densities, component density, are plotted (so that the histogram has a total area of one).
 Defaults to TRUE if and only if breaks are equidistant (and probability is not specified).}

\item{image_width}{Width of image to be saved}

\item{image_height}{Height of image to be saved}
}
\value{
Describe output
}
\description{
Plot and Save Histograms of nearest neighbor distances (\code{\link[spatstat]{nndist}})
}
\details{
Takes a point pattern as input (or two if using multitype \code{\link[spatstat]{nncross}}) and returns the histograms of all nearest neighbors
in the nn_vec argument.  If `output` is set to "save", then it will also save the image.
}
