% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_02.R
\name{nndist_subset}
\alias{nndist_subset}
\title{nndist_subset}
\usage{
nndist_subset(
  pattern,
  pattern2 = NULL,
  window = NULL,
  drop_isolated = FALSE,
  k = 1,
  output = "list"
)
}
\arguments{
\item{pattern}{point pattern of type \emph{ppp} or \emph{pp3} to calculate distances from}

\item{pattern2}{point pattern of type \emph{ppp} or \emph{pp3} to calculate distances to.
If \emph{NULL} then \emph{pattern} will be used}

\item{window}{object of class \emph{owin} (for ppp) or \emph{box3} (for pp3).
Only points inside the window are used to calculate distances from, but points outside the
window are still included as potential neighbors. If \emph{NULL} then it includes the
entire domain of \emph{pattern}}

\item{drop_isolated}{if \emph{TRUE} then points that are closer to a boundary than they
are to their \emph{kth} NN will be dropped: their distances are set to \emph{NA}}

\item{k}{an integer or vector of integers that determines which NN's to calculate}

\item{output}{outputs a list if "list", outputs a matrix otherwise}
}
\description{
Edge corrected nearest neighbor distribution
}
\details{
This function calculates the distances to the \emph{kth} nearest neighbors (NN's) for
a subset of points, defined as those inside of \emph{window}. All points are still considered
when fiding the NN's.
}
