% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp_functions.R
\name{nnR}
\alias{nnR}
\title{Find Nearest Neighbors within a given radius}
\usage{
nnR(X, r)
}
\arguments{
\item{X}{\code{\link[spatstat]{pp3}} object to find nearest points in.}

\item{r}{Radius to find nearest neighbors within.}
}
\value{
Returns a list of:
[[1]] The nearest neighbor distances.
[[2]] The nearest neighbor indixes.
}
\description{
Finds all nearest neighbors within a given input radius of each point
in the pattern. This function is not needed and slow.
\code{\link[spatstat]{closepairs.pp3}} will do the same thing but more
efficiently.
}
