% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_functions.R
\name{panomK3est}
\alias{panomK3est}
\title{Perform anomaly K3est envelope calculations}
\usage{
panomK3est(perc, pattern, nEvals, rmax = NULL, nrval = 128,
  correction = "iso", toSub = NULL)
}
\arguments{
\item{perc}{The fraction of points to select randomly each time out of the
original \code{\link[spatstat]{pp3}} object. Number between 0 and 1.}

\item{pattern}{The original \code{\link[spatstat]{pp3}} object.}

\item{nEvals}{The number of random relabelings and  that should be performed.}

\item{rmax}{See \code{\link[spatstat]{K3est}}. Maximum radius to be
calculated for \code{\link[spatstat]{K3est}}.}

\item{nrval}{See \code{\link[spatstat]{K3est}}. Number of radii that
\code{\link[spatstat]{K3est}} should be calculated at.}

\item{toSub}{If NULL, use the 50th percentile of the calculated set of
\code{\link[spatstat]{K3est}} envelopes to subtract off. Otherwise, the
second, [[2]], entry in the list returned from this same function. This is
how to compare envelope calculations from different point patterns. You
must subtract the same values from both data sets. toSub allows you to
input the values that were subtracted from a previous set of envelopes, for
comparison.}

\item{iso}{Either "iso", "trans", or "bord" edge correction.}
}
\value{
A list of: [[1]] Matrix of data for all relabelings. Can be plotted
  using \code{\link{envPlot}}. [[2]] Vector containing the values that were
  subtracted from the results at each r value. Can be used to subtract from
  another set of envelopes for comparison. [[3]] rmax used in the
  calculation. [[4]] nrval used in the calculation.
}
\description{
See \code{\link{pK3est}}. Performs exactly the same as this function, except
returns the "anomaly K3est" results. This means that it returns the square
room of the \code{\link[spatstat]{K3est}} results, with the 50th percentile
subtracted out. This centers envelopes around zero, and the square root
standardized variance across all r values. See book at
\url{http://spatstat.org/book.html} for a good statistical reference.
}
\section{Edge Corrections}{
 See \code{\link[spatstat]{Kest}} or book availible
  at \url{http://spatstat.org/book.html} for more info on these edge
  corrections.

\subsection{Isotropic - "iso"}{Isotropic edge correction. Assumes point
pattern is isotropic, or that it can rotate in space without changing
statistics.}
\subsection{Translation - "trans"}{Translation edge correction.
Assumes translation of point pattern does not change statistics.}
\subsection{Border - "bord"}{Border edge correction. Makes no assumptions
about data. Uses only data provided in the original point pattern. Only
evaluates \code{\link[spatstat]{K3est}} when the radius of the search stays
within the domain of the point pattern itself.}
}

