% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_recon.R
\name{paraOrthoY}
\alias{paraOrthoY}
\title{Internal function used by \code{\link{paraOrthoRec}}.}
\usage{
paraOrthoY(x, y, a, d)
}
\description{
Internal function used by \code{\link{paraOrthoRec}}.
}
\seealso{
\code{\link{paraOrthoRec}}
}
