% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-recon.R
\name{paraRec}
\alias{paraRec}
\title{Reconstruction using a parabolic tip shape assumption.}
\usage{
paraRec(ato, a, d = 9e+08, icf = 1)
}
\description{
Reconstruction using a parabolic tip shape assumption.
}
