% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_envelopes.R
\name{pEnvelope}
\alias{pEnvelope}
\title{Create an Envelope in Parallel}
\usage{
pEnvelope(cl, X, fun = K3est, nsim = 99, nrank = 1, ..., simulate = NULL)
}
\arguments{
\item{cl}{A "cluster" object (\emph{e.g.} as generated by
\code{\link[parallel]{makeCluster}})}

\item{X}{A point pattern (object of class "pp3")}

\item{fun}{Function that computes the desired summary statistic for a 3D
point pattern.}

\item{nsim}{Number of simulated point patterns to be generated when computing
the envelopes.}

\item{nrank}{Integer. Rank of the envelope value amongst the \code{nsim} simulated
values. A rank of 1 means that the minimum and maximum simulated values
will be used.}

\item{...}{Extra arguments passed to \code{fun}}

\item{simulate}{Optional. Specifies how to generate the simulated point
patterns. If \code{simulate} is an expression in the R language, then this
expression will be evaluated \code{nsim} times, to obtain \code{nsim} point patterns
which are taken as the simulated patterns from which the envelopes are
computed. If \code{simulate} is a function, then this function will be
repeatedly applied to the point pattern \code{X} to obtain \code{nsim} simulated
patterns. If \code{simulate} is a list of point patterns, then the entries in
this list will be treated as the simulated patterns from which the
envelopes are computed. \strong{Currently, only a list of point patterns is
accepted.}}
}
\value{
A function value table (object of class "fv") which can be plotted
directly. See \code{\link[spatstat]{envelope}} for further details.
}
\description{
Create an Envelope in Parallel
}
\seealso{
\link[spatstat]{envelope}, \link[spatstat]{K3est},
\link[spatstat]{G3est}, \link[spatstat]{F3est}, \link[spatstat]{pcf3est}
}
