% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_envelopes.R
\name{pF3est}
\alias{pF3est}
\title{Perform F3est on random relabelings in parallel}
\usage{
pF3est(perc, pattern, nEvals, rmax = NULL, nrval = 128,
  correction = "rs", anom = FALSE, toSub = NULL, cores = NULL)
}
\arguments{
\item{perc}{The fraction of points to select randomly each time out of the
original \code{\link[spatstat]{pp3}} object. Number between 0 and 1.}

\item{pattern}{The original \code{\link[spatstat]{pp3}} object.}

\item{nEvals}{The number of random relabelings and  that should be performed.}

\item{rmax}{See \code{\link[spatstat]{F3est}}. Maximum radius to be
calculated for \code{\link[spatstat]{F3est}}.}

\item{nrval}{See \code{\link[spatstat]{F3est}}. Number of radii that
\code{\link[spatstat]{F3est}} should be calculated at.}

\item{correction}{Either "rs", "km", or "cs" edge correction.}

\item{anom}{Whether or not to retun the anomaly results. \code{TRUE} or
\code{FALSE}. See section below for more info.}

\item{toSub}{The numeric vector of data to subtract for the "anom" pF3est.
Only used when \code{anom = TRUE}. See below for more info.}

\item{cores}{Number of cores to use for parallel processing. If no parallel
desired, set to 1. If NULL, will automatically set to the number of cores
available on the machine.}
}
\value{
\subsection{For \code{anom = FALSE}}{Returns a matrix containing the
data from the all of the \code{\link[spatstat]{F3est}} runs on different
re-labelings. Can plot data using \code{\link{envPlot}}.} \subsection{For
\code{anom = TRUE}}{A list of: [[1]] Matrix of data for all relabelings. Can
be plotted using \code{\link{envPlot}}. [[2]] Vector containing the values
that were subtracted from the results at each r value. Can be used to
subtract from another set of envelopes for comparison. [[3]] rmax used in the
calculation. [[4]] nrval used in the calculation.}
}
\description{
\code{pF3est} first randomly relabels a specified percentage of points from
the original \code{\link[spatstat]{pp3}} object. It then performs a 3D F
function test (\code{\link[spatstat]{F3est}}) on these selected points. It
repeats this as many times as specified. These tests are run in parallel to
increase computation speed.
}
\section{Edge Corrections}{
 See  book availible at
  \url{http://spatstat.org/book.html} for more info on these edge
  corrections.

  \subsection{Reduced Sample - "rs"}{} \subsection{Kaplan-Meier - "km"}{}
  \subsection{Chiu-Stoyan (aka Hanisch) - "cs"}{}
}

\section{Anomaly F3est}{
 When \code{anom = TRUE}, the function returns the
  anomaly \code{\link[spatstat]{F3est}}.This means that it returns the
  \code{\link[spatstat]{F3est}} results with the 50th percentile subtracted
  out. This centers envelopes around zero.

  \code{toSub} is an argumet to be paired with \code{anom = TRUE}. If NULL,
  use the 50th percentile of the calculated set of
  \code{\link[spatstat]{F3est}} envelopes to subtract off. Otherwise, use the
  second, [[2]], entry in the list returned from this same function. This is
  how to compare envelope calculations from different point patterns. You
  must subtract the same values from both data sets. toSub allows you to
  input the values that were subtracted from a previous set of envelopes, for
  comparison.
}

