% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_functions.R
\name{pK3est}
\alias{pK3est}
\title{Perform K3est on random relabelings in parallel}
\usage{
pK3est(perc, pattern, nEvals, rmax = NULL, nrval = 128,
  correction = "iso")
}
\arguments{
\item{perc}{The fraction of points to select randomly each time out of the
original \code{\link[spatstat]{pp3}} object. Number between 0 and 1.}

\item{pattern}{The original \code{\link[spatstat]{pp3}} object.}

\item{nEvals}{The number of random relabelings and  that should be performed.}

\item{rmax}{See \code{\link[spatstat]{K3est}}. Maximum radius to be
calculated for \code{\link[spatstat]{K3est}}.}

\item{nrval}{See \code{\link[spatstat]{K3est}}. Number of radii that
\code{\link[spatstat]{K3est}} should be calculated at.}

\item{iso}{Either "iso", "trans", or "bord" edge correction.}
}
\value{
Returns a matrix containing the data from all of the
  \code{\link[spatstat]{K3est}} runs on different re-labelings. Can plot data
  using \code{\link{envPlot}}.
}
\description{
\code{pK3est} first randomly relabels a specified percentage of points from
the original \code{\link[spatstat]{pp3}} object. It then performs a 3D K
function test (\code{\link[spatstat]{K3est}}) on these selected points. It
repeats this as many times as specified. These tests are run in parallel to
increase computation speed.
}
\section{Edge Corrections}{
 See \code{\link[spatstat]{Kest}} or book availible
  at \url{http://spatstat.org/book.html} for more info on these edge
  corrections.

\subsection{Isotropic - "iso"}{Isotropic edge correction. Assumes point
pattern is isotropic, or that it can rotate in space without changing
statistics.}
\subsection{Translation - "trans"}{Translation edge correction.
Assumes translation of point pattern does not change statistics.}
\subsection{Border - "bord"}{Border edge correction. Makes no assumptions
about data. Uses only data provided in the original point pattern. Only
evaluates \code{\link[spatstat]{K3est}} when the radius of the search stays
within the domain of the point pattern itself.}
}

