% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{plot3d.pp3}
\alias{plot3d.pp3}
\title{Plot a \code{\link[spatstat]{pp3}} in a manipulatable 3D plot.}
\usage{
\method{plot3d}{pp3}(X, ...)
}
\arguments{
\item{X}{A \code{pp3}. The point pattern to visualize}

\item{...}{Other arguments to pass to \code{plot3d} from the \code{rgl}
library.}
}
\description{
(requires the rgl library)
}
\seealso{
\code{\link[rgl]{plot3d}}
}
