% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_summarize.R
\name{plot_summary_dev}
\alias{plot_summary_dev}
\title{Plot summary functions (developement version)}
\usage{
plot_summary_dev(
  func = "K",
  observed_values,
  envelopes,
  ...,
  pattern.colors = c(`95.0\% AI` = "pink", Median = "black", Observed = "blue"),
  fill.colors = pattern.colors,
  base_value = "first",
  unit = "nm",
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km",
  legend.key.size = 20,
  legend.text.size = 20,
  legend.position = c(0.75, 0.8),
  axis.title.size = 40,
  title = "",
  title.size = 40,
  axis.text.x.size = 40,
  axis.text.y.size = 40,
  alpha = 0.5,
  linewidth = 0.5,
  env_linewidth = 0.5,
  linetype = "solid",
  env_linetype = "dashed"
)
}
\arguments{
\item{func}{Summary function to plot}

\item{observed_values}{a list. observed summary function value}

\item{envelopes}{a list theoretical values found by function such as \verb{calc_summary_funcs().} Should be list}

\item{pattern.colors}{colors and names for envelope lines}

\item{fill.colors}{colors and names for envelope fill.  Match names to \code{pattern.colors}}

\item{K_cor}{edge correction(s) to be used for K function}

\item{G_cor}{edge correction(s) to be used for G function}

\item{F_cor}{edge correction(s) to be used for F function}

\item{GXGH_cor}{edge correction(s) to be used for GXGH function}

\item{GXHG_cor}{edge correction(s) to be used for GXHG function}
}
\description{
Under developement. Plot the observed value with envelopes for expected values for summary function
}
