% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{prettyPlot}
\alias{prettyPlot}
\title{Plot a pretty \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}.}
\usage{
prettyPlot(
  ms,
  rng = NA,
  xlim = NULL,
  ylim = NULL,
  lwd = 1,
  main = NULL,
  hold.par = F,
  ...
)
}
\arguments{
\item{ms}{The \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} to
be plotted}

\item{rng}{Optional. The RRNG \code{data.frame} used to color ranged peaks in the
mass spectrum}

\item{hold.par}{}
}
\description{
\code{prettyPlot}
}
