% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_envelopes.R
\name{pRlabel}
\alias{pRlabel}
\title{Relabel a Point Pattern in Parallel}
\usage{
pRlabel(cl, X, labels = marks(X), permute = TRUE, nsim = 99)
}
\arguments{
\item{X}{Point pattern (object of class "ppp", "lpp", "pp3" or "ppx")
or line segment pattern (object of class "psp")}

\item{labels}{Vector of values from which the new marks will be drawn at
random. Defaults to the vector of existing marks}

\item{permute}{Logical value indicating whether to generate new marks by
randomly permuting labels or by drawing a random sample with replacement}

\item{nsim}{Number of simulated realisations to be generated. Default is 99}
}
\value{
A list of marked point patterns, each of the same class of X.
}
\description{
Randomly allocates marks to a point pattern, or permutes the existing marks,
or resamples from the existing marks.
}
\seealso{
\code{\link[spatstat]{rlabel}}
}
