% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{quadrats.pp3}
\alias{quadrats.pp3}
\title{Extension of \code{\link[spatstat]{quadrats}} to \code{\link[spatstat]{pp3}} objects.}
\usage{
quadrats.pp3(X, nx = 5, ny = 5, nz = 5)
}
\arguments{
\item{X}{The \code{\link[spatstat]{pp3}} object to split up.}

\item{nx, ny, nz}{Number of ractangular quadrats in the x, y, and z directions.}
}
\value{
A list containing the split up \code{pp3} objects.
}
\description{
Divides volume into quadrats and returns them..
}
