% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{randomTakeAway}
\alias{randomTakeAway}
\title{Helper for \code{\link{makecluster}} to take away random cluster points}
\usage{
randomTakeAway(cluster.Indices, n, N)
}
\arguments{
\item{cluster.Indices}{A vector containing the indices of the current cluster
points}

\item{n}{The number of points that need to be removed randomly}

\item{N}{the number of points in the entire underlaying pattern.}
}
\value{
New indices vector containing new cluster points randomly placed.
}
\description{
When \code{superfast} argument of \code{\link{makecluster}} is \code{TRUE},
there may be random points that need to be removed from the existing clusters
of the the underlaying pattern. This function does just that.
}
