% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{rangeCount}
\alias{rangeCount}
\title{Count the hits within a mass range}
\usage{
rangeCount(pos, start, end)
}
\description{
To get the hit information within a mass range, use \code{\link{rangePOS}}
}
\seealso{
\code{\link{rngCount}}, \code{\link{rangePOS}}
}
