% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{rangePeaks}
\alias{rangePeaks}
\title{Automatically define ranges from a \code{\link[MALDIquant]{MassPeaks}}
object.}
\usage{
rangePeaks(spec, peaks, cut = 0.1)
}
\description{
Automatically define ranges from a \code{\link[MALDIquant]{MassPeaks}}
object.
}
