% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{rangePOS}
\alias{rangePOS}
\title{Extract hits within a given mass range.}
\usage{
rangePOS(pos, start, end)
}
\description{
rangePOS extracts the rows of a \code{POS} or \code{ATO} object whose mass
is within the provided range.
}
