% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{rangeSpec}
\alias{rangeSpec}
\title{Define ranges in a \code{\link[MALDIquant]{MassSpectrum}}.}
\usage{
rangeSpec(cform, charge = 1:4, resolution = 400, thresh = 20)
}
\description{
Define ranges in a \code{\link[MALDIquant]{MassSpectrum}}.
}
