% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{rcp.simple.relab}
\alias{rcp.simple.relab}
\title{Perform num.relabs relabelings}
\usage{
rcp.simple.relab(
  exp.ppp,
  num.relabs,
  thickness,
  rcp.list,
  maxGr,
  maxKr,
  nGr,
  nKr,
  ncores = detectCores()
)
}
\description{
Perform num.relabs relabelings
}
