% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{rcp_simple_relab}
\alias{rcp_simple_relab}
\title{Perform num_relabs relabelings}
\usage{
rcp_simple_relab(
  exp_ppp,
  num_relabs,
  thickness,
  neighbor_number = 1,
  rcp_list,
  maxGr,
  maxKr,
  nGr,
  nKr,
  ncores = detectCores()
)
}
\description{
Perform num_relabs relabelings
}
