% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_depracated.R
\name{rcp_simple_relab_v01}
\alias{rcp_simple_relab_v01}
\title{Perform num_relabs relabelings
DEPRACATED now use relabler}
\usage{
rcp_simple_relab_v01(
  exp_ppp,
  num_relabs,
  thickness,
  neighbor_number = 1,
  rcp_list,
  maxGr,
  maxKr,
  nGr,
  nKr,
  ncores = detectCores()
)
}
\description{
Perform num_relabs relabelings
DEPRACATED now use relabler
}
