% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_rcp-sim.R
\name{read.rcp}
\alias{read.rcp}
\title{Read an RCP Simulation}
\usage{
read.rcp(fpath_config, fpath_sys, scaleUp = FALSE, newRadius = 0.5)
}
\arguments{
\item{fpath_config}{The file path to the RCP FinalConfig file.}

\item{fpath_sys}{The file path to the associated RCP system file.}

\item{newRadius}{Numeric. If \code{scaleUP = TRUE}, this is the new radius
that the RCP particles will be scaled to. Default is 0.5.}

\item{scaleUP}{Boolean. If \code{TRUE}, scales RCP so particles have new
radius. If \code{FALSE}, RCP stays as generated.}
}
\value{
A \code{\link[spatstat]{pp3}} object of the RCP pattern.
}
\description{
Reads in a RCP Generator output file as a \code{\link[spatstat]{pp3}} object.
}
\details{
The RCP generation is described in Desmond & Weeks,
  "Random Close Packing of Disks and Spheres in Confined Geometries",
  \emph{Physical Review E}, \strong{80} (5), 051305 (2009):
  \url{https://doi.org/10.1103/PhysRevE.80.051305}; the algorithm can be
  downloaded at
  \url{http://www.physics.emory.edu/faculty/weeks/ken/RCPAlgorithm.html}.
}
\references{
Desmond & Weeks,
  "Random Close Packing of Disks and Spheres in Confined Geometries",
  \emph{Physical Review E}, \strong{80} (5), 051305 (2009):
  \url{https://doi.org/10.1103/PhysRevE.80.051305}
}
\seealso{
\href{http://www.physics.emory.edu/faculty/weeks/ken/RCPAlgorithm.html}{RCP
Generator Algorithm}

Other simulation functions: 
\code{\link{scaleRCP}()},
\code{\link{stitch.size}()}
}
\concept{simulation functions}
