% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{readPOS}
\alias{readPOS}
\title{Read a POS file.}
\usage{
readPOS(filepath)
}
\arguments{
\item{filepath}{A string. The file path to the POS file.}
}
\value{
A data.frame with columns corresponding to the x, y, and z positions
  of the reconstruction and the mass-to-charge ratio.
}
\description{
\code{readPOS} reads a POS file (from IVAS) into a data frame.
}
\references{
Local Electrode Atom Probe Tomography: A User's Guide
}
\seealso{
\code{\link{readATO}}
}
