% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{readResult}
\alias{readResult}
\title{Read TAPSim Results}
\usage{
readResult(fp, type = "ASCII")
}
\arguments{
\item{fp}{Character. The filepath}

\item{type}{Character. The file type to be read. Currently, only ASCII is
implemented.}
}
\value{
A \code{data.frame} with the structure of the TAPSim output.
}
\description{
Given a filepath, read in a TAPSim results file.
\emph{Currently only implemented for ASCII results files.}
}
\seealso{
\url{
  http://www.uni-stuttgart.de/imw/mp/forschung/atom_probe_RD_center/howto.pdf
}

\url{https://d-nb.info/1138282715/34}

Other TAPSim functions: 
\code{\link{resultToDet}()},
\code{\link{resultToPOS}()}
}
\concept{TAPSim functions}
