% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{readRRNG}
\alias{readRRNG}
\title{Read an RRNG file (from IVAS) into a dataframe}
\usage{
readRRNG(filepath)
}
\arguments{
\item{filepath}{Character. A filepath to the RRNG file.}
}
\value{
A \code{data.frame} with columns corresponding to the mass ranges,
  ion volume, name, molecular formula (\code{NA} if not present), and color
  for display.
}
\description{
\code{readRRNG} reads an RRNG file from IVAS into an RNG object for ranging
a \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} generated by
\code{\link{createSpec}}
}
\references{
Larson \emph{et al.},
  \emph{Local Electrode Atom Probe Tomography: A User's Guide} (2009)
}
\seealso{
Other APT import functions: 
\code{\link{readATO}()},
\code{\link{readPOS}()}
}
\concept{APT import functions}
