% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-tapsim.R
\name{readTAPSim}
\alias{readTAPSim}
\title{Read a TAPSim results file}
\usage{
readTAPSim(fp, type = "results")
}
\arguments{
\item{fp}{The filepath}

\item{type}{The file type to be read. Currently, only "results" is
implemented.}
}
\value{
A data.frame
}
\description{
Given a filepath, read in a TAPSim file. Currently only implemented for
results.
}
\seealso{
\url{
  http://www.uni-stuttgart.de/imw/mp/forschung/atom_probe_RD_center/howto.pdf
}
}
