% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{refSpec}
\alias{refSpec}
\title{Create a reference \code{\link[MALDIquant]{MassPeaks}} object for warping.}
\usage{
refSpec(form, charge = 1:4, resolution = 400)
}
\description{
Create a reference \code{\link[MALDIquant]{MassPeaks}} object for warping.
}
