% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_summarize.R
\name{relabel_summarize}
\alias{relabel_summarize}
\title{Relabel input point pattern and calcluate summary functions}
\usage{
relabel_summarize(
  seed,
  pattern,
  funcs = c("K", "G", "F", "GXGH"),
  dopant_formula,
  host_formula,
  k = 1,
  maxKr = 10,
  nKr = 200,
  maxGr = 5,
  nGr = 1000,
  maxGXGHr = 3,
  maxGXHGr = 8,
  nGXr = 1000,
  vside = 0.3,
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{seed}{number to use in \code{set.seed} for reproducibility}

\item{pattern}{point pattern of type ppp or pp3}

\item{funcs}{vector of summary functions to calculate}

\item{dopant_formula}{mark of points from which to calculate summary functions}

\item{host_formula}{marks of points to not use in summary functions (except cross type functions)}

\item{K_cor}{edge correction(s) to be used for K function}

\item{G_cor}{edge correction(s) to be used for G function}

\item{F_cor}{edge correction(s) to be used for F function}

\item{GXGH_cor}{edge correction(s) to be used for GXGH function}

\item{GXHG_cor}{edge correction(s) to be used for GXHG function}

\item{...}{maxKr, nKr, maxGr, nGr, maxGXGHr, maxGXHGr, nGXr}
}
\value{
summary functions listed in \code{funcs} variable
}
\description{
takes an input binary point pattern, randomly relabels it while maintainig the
same ratio of marks, and calculates summary functions.  Can then use  \code{\link[rapt]{average_relabelings}}
to find averages and envelopes
}
