% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_summarize.R
\name{rescale_pattern}
\alias{rescale_pattern}
\title{rescale_pattern Function}
\usage{
rescale_pattern(pattern, new_intensity)
}
\arguments{
\item{pattern}{3D point pattern (pp3 object) to be scaled}

\item{new_intensity}{final intensity of point pattern}
}
\value{
pp3_new will preserve all properties of the input \emph{pattern}, except the intensity
}
\description{
Rescales an input 3D point pattern
}
\details{
Takes an input 3D point pattern (pp3 object) and scales it to have intensity of new.intensity by multiplying
the coordinates of each point as well as the window.
}
