% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{resultToDet}
\alias{resultToDet}
\title{Convert a TAPSim results file to a detector point pattern}
\usage{
resultToDet(res, clip.radius = NULL)
}
\arguments{
\item{res}{The results data.frame}

\item{clip.radius}{The detector radius at which to clip the points}
}
\value{
A ppp containing the detector positions marked by their id
}
\description{
Convert a TAPSim results file to a detector point pattern
}
\seealso{
\code{\link{createDet}}, \code{\link{readResult}}
}
