% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{resultToPOS}
\alias{resultToPOS}
\title{Convert a TAPSim results file to a POS}
\usage{
resultToPOS(res, clip.radius = NULL)
}
\arguments{
\item{res}{The results data.frame}

\item{clip.radius}{The detector radius at which to clip the points}
}
\value{
A data.frame containing the (x,y,z) position and "mass' (id) of the
  point in the stame structure of a POS as created by \code{\link{readPOS}}.
}
\description{
Convert a TAPSim results file to a POS
}
\seealso{
\code{\link{readPOS}}, \code{\link{readResult}}
}
