% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{rjitter.ppp}
\alias{rjitter.ppp}
\title{Random Perturbation of a Point Pattern}
\usage{
rjitter.ppp(
  X,
  radius,
  retry = TRUE,
  giveup = 10000,
  ...,
  nsim = 1,
  drop = TRUE
)
}
\description{
Random Perturbation of a Point Pattern
}
\seealso{
\code{\link[spatstat]{rjitter}}, \code{\link{rjitter.pp3}}
}
