% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-sim.R
\name{rjitter3}
\alias{rjitter3}
\title{Extends the \code{\link[spatstat]{rjitter}} to \code{\link[spatstat]{pp3}}.}
\usage{
rjitter3(X, domain = box3())
}
\description{
Extends the \code{\link[spatstat]{rjitter}} to \code{\link[spatstat]{pp3}}.
}
