% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{rownames.pp3<-}
\alias{rownames.pp3<-}
\title{Extends \code{\link[base:row+colnames]{rownames}} to
\code{\link[spatstat]{pp3}}.}
\usage{
rownames.pp3(pat, lab) <- value
}
\arguments{
\item{pat}{\code{pp3}. The point pattern to assign rownames.}

\item{lab}{character. The new rownames.}
}
\description{
Extends \code{\link[base:row+colnames]{rownames}} to
\code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[base:row+colnames]{rownames}}
}
