% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-extend.R
\name{rownames.pp3}
\alias{rownames.pp3}
\title{Extends \code{\link[base:row+colnames]{rownames}} to \code{\link[spatstat]{pp3}}.}
\usage{
rownames.pp3(pat)
}
\arguments{
\item{pat}{A \code{pp3}. The point pattern from which to extract rownames.}
}
\value{
A string vector. The rownames of the point pattern.
}
\description{
Extends \code{\link[base:row+colnames]{rownames}} to \code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[base:row+colnames]{rownames}}
}
