% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim.R
\name{rpoint3}
\alias{rpoint3}
\title{Generate N Random Points}
\usage{
rpoint3(
  n,
  f,
  fmax = 1,
  win = box3(),
  ...,
  giveup = 1000,
  verbose = FALSE,
  nsim = 1,
  drop = TRUE
)
}
\description{
\code{rpoint3} extends \code{\link[spatstat]{rpoint}} to
\code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[spatstat]{rpoint}}
}
