% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim.R
\name{rPoissonCluster3}
\alias{rPoissonCluster3}
\title{Simulate 3D Poisson Cluster Process}
\usage{
rPoissonCluster3(
  kappa,
  expand,
  rcluster,
  win = box3(),
  ...,
  nsim = 1,
  drop = T
)
}
\description{
\code{rPoissonCluster3} extends \code{\link[spatstat]{rPoissonCluster}} to
\code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[spatstat]{rpoint}}
}
