% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{rrl_averager}
\alias{rrl_averager}
\title{Functionto take all relabelings and return averages and envelope values}
\usage{
rrl_averager(
  relabelings,
  envelope.value = 0.95,
  funcs = c("K", "G"),
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{relabelings}{output of `relabeler` function}

\item{envelope.value}{size of envelope to compute.  Should be decimal (e.g. 0.95 = 95%)}
}
\description{
Functionto take all relabelings and return averages and envelope values
}
