% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-extend.R
\name{sample.pp3}
\alias{sample.pp3}
\title{Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{pp3}}.}
\usage{
sample.pp3(X, size)
}
\description{
Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{pp3}}.
}
