% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_multi.R
\name{saxeyCor}
\alias{saxeyCor}
\title{Create a Corrected 2D correlation histogram}
\usage{
saxeyCor(
  ind,
  ato,
  begin,
  end,
  res = 0.5,
  plot.it = TRUE,
  expected = "DOUBLE",
  log = FALSE,
  remove_diagonal = FALSE,
  remove_double = FALSE,
  remove_diagonal_width = 0,
  remove_double_width = 0,
  expected_remove_threshold = NA,
  ...
)
}
\arguments{
\item{ind}{A vector of first hit indices for double hits.  The list
generated by \code{\link{indexMultiples}} has this vector as its second entry.}

\item{ato}{An ATO data frame.  This data frame is generated by \code{\link{readATO}}.}

\item{begin}{A numeric.  The starting mass value of the correlation histogram}

\item{end}{A numeric.  The ending mass value of the correlation histogram}

\item{plot.it}{A logical scalar.  Should the result be plotted?}

\item{expected}{either \code{"DOUBLE"}, \code{"TOTAL"}, or \code{FALSE}.  Determines which method of normalization to use. If \code{"DOUBLE"} then
e_ij and N are calculated using double hits.  If \code{"TOTAL"} then e_ij and N are calculated using all hits.  If \code{FALSE},
no normalization is done and simply pair hits, p_ij, are returned}

\item{log}{a logical.  Should the log of the results be plotted?}

\item{remove_diagonal}{a logical.  Should the diagonal  (y = x) be removed?  This effectively removes
values of p_ii/d_ii}

\item{remove_double}{a logical.  Should the "double diagonal" lines (y = 2x and y = 1/2x) be removed?}

\item{remove_double_width}{an integer.  If \code{remove_double_width} is true, this will
remove \code{remove_double_width} additional pixels from each side of the double (y = 2x, y = 1/2x) lines}

\item{expected_remove_threshold}{A numeric.  Will remove all pixels whose expectation values e_ij are less than
this threshold}

\item{...}{extra input to be added to \code{\link{image.plot}} function}

\item{res.}{A numeric.  The bin width of the correlation histogram}

\item{remove_diagona_width}{an integer.  If \code{remove_diagonal} is true, this will
remove \code{remove_diagonal_width} additional pixels from each side of the diagonal line,
to account for lower resolution in the mass spectrum}
}
\description{
\code{saxeyCor} creates a 2D correlation histogram, similar to that of \code{saxeyPlot}
However, \code{saxeyCor} differs in that it offers various normalization options.  This is the
"correlation table" referred to in section 2.1 of Saxey, D. W. (2011).
}
\details{
This is used to reproduce plots made in of Saxey, D. W. (2011). If \code{expected = FALSE},
then the uncorrected correlation histograms shown in Figure 3 of Saxey, D. W. (2011), are produced.
If \code{expected = "DOUBLE"}, then the calculations will follow equations 2.1-2.3 and d_ij will be plotted.
#' If \code{expected = "TOTAL"}, then the calculations will follow equations 2.1-2.3, except
expectation values e_ij are calculated using all hits, not just double hits
}
\references{
Saxey, D. W. (2011). Correlated ion analysis and the
interpretation of atom probe mass spectra. Ultramicroscopy, 111(6), 473-479.
\url{https://doi.org/10.1016/j.ultramic.2010.11.021}.
}
\seealso{
\code{\link{indexMultiples}}, \code{\link[ash]{ash2}}
}
