% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_multi.R
\name{saxeyPlot}
\alias{saxeyPlot}
\title{Create a 2D correlation histogram}
\usage{
saxeyPlot(ind, ato, begin, end, res = 0.25, plot.it = T)
}
\arguments{
\item{ind}{A vector of first hit indices for double hits. The list
generated by \code{\link{indexMultiples}} has this vector as its second
entry.}

\item{ato}{An ATO data frame. This data frame is generated by
\code{\link{readATO}}.}

\item{begin}{A numeric. The starting mass value of the correlation histogram.}

\item{end}{A numeric. The ending mass value of the correlation histogram.}

\item{res}{A numeric. The bin width of the correlation histogram.}

\item{plot.it}{A logical scalar. Should the result be plotted?}
}
\value{
The output is the same as that of \code{\link[ash]{ash2}}.
}
\description{
\code{saxeyPlot} creates a 2D correlation histogram of mass spectra using the
index of the first hit of an ATO data frame. The optional plot also shows the
corresponding mass spectrum on each axis for easier identification of high
intensity areas of the histogram.
}
\references{
Saxey, D. W. (2011). Correlated ion analysis and the
interpretation of atom probe mass spectra. Ultramicroscopy, 111(6), 473-479.
\url{https://doi.org/10.1016/j.ultramic.2010.11.021}.
}
\seealso{
\code{\link{indexMultiples}}, \code{\link[ash]{ash2}}
}
