% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim-rcp.R
\name{scale}
\alias{scale}
\title{Scale an RCP pattern}
\usage{
scale(pp3file, newRadius = 0.5, oldRadius = NULL, win = NULL)
}
\arguments{
\item{pp3file}{A \code{pp3}. A point pattern to scale.}

\item{newRadius}{A numeric. The radius that you would like your particles
to have after scaling.}

\item{oldRadius}{A numeric. The radius that your particles have in the
original \code{pp3}.}

\item{win}{A \code{box3}. A domain indicating the size of the rcp
generation (leave null if integer values, the program will figure it out).}
}
\description{
\code{scale} scales an RCP file to a desired radius, given the original
radius.
}
