% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{shift.pp3}
\alias{shift.pp3}
\title{Extends \code{\link[spatstat]{shift}} to \code{\link[spatstat]{pp3}}.}
\usage{
\method{shift}{pp3}(X, vec = c(0, 0, 0), ..., origin = NULL)
}
\description{
Extends \code{\link[spatstat]{shift}} to \code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[spatstat]{shift}}

Other spatstat extensions: 
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
