% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_recon.R
\name{sphereRec}
\alias{sphereRec}
\title{Reconstruction using the sphere on cone method}
\usage{
sphereRec(ato, r, d = 9e+08, icf = 1)
}
\description{
\code{sphereRec} creates a reconstruction based on a sphere on post tip
shape assumption. A conical post will be implemented in the future.
}
