% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim-rcp.R, R/rcp_functions.R
\name{stitch}
\alias{stitch}
\title{Stitch a \code{pp3} with periodic boundary conditions.}
\usage{
stitch(pp3file, reps = c(2, 2, 2), win = NULL)

stitch(pp3file, reps = c(2, 2, 2), win = NULL)
}
\arguments{
\item{pp3file}{A \code{pp3}. A point pattern with periodic boundary
conditions which you want to stitch together.}

\item{reps}{A numeric vector. The number of iterations you want in each
direction: c(xreps, yreps, zreps).}

\item{win}{A \code{box3}. The size of the rcp generation (leave null if
integer values, the program will figure it out).}

\item{pp3file}{A \code{\link[spatstat]{pp3}} object containing the RCP
generated 3D point positions.}

\item{reps}{A numerical vector containing the number of repetitions you would
like in the x, y and z directions, respectively. c(x,y,z).}

\item{win}{A \code{\link[spatstat]{box3}} object indicating the size of the
original rcp generation. Can leave blank if these are integer value.}
}
\value{
Will return a \code{\link[spatstat]{pp3}} object with the total
  stiched point pattern.
}
\description{
\code{stitch} takes a \code{pp3} generated with periodic boundary conditions
(usually read in from an external RCP generator) and tiles it together.

Takes RCP point patterns with periodic boundary conditions and creates a
single, larger point cloud by stitching these patterns together.
}
