% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcp_functions.R
\name{stitch.size}
\alias{stitch.size}
\title{Stitches together a RCP point pattern with periodic boundary conditions}
\usage{
stitch.size(pp3file, win = NULL, boxSize)
}
\arguments{
\item{pp3file}{A \code{\link[spatstat]{pp3}} object containing the RCP
generated 3D point positions.}

\item{win}{A \code{\link[spatstat]{box3}} object indicating the size of the
original rcp generation. Can leave blank if these are integer value.}

\item{boxSize}{A numeric vector of the dimensions that you want in the final
\code{\link[spatstat]{pp3}} object: c(xmax,ymax,zmax). Assumes that
(xmin,ymin,zmin) = (0,0,0).}
}
\description{
Similar to \code{\link{stitch}}. Instead of only inputting the number of
repetitions in each dimension, \code{stitch.size} allows you to specify the
domain size that you want to return, even if it is not an integer multiple of
the original dimensions.
}
