% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{studpermu.test.hyperframe}
\alias{studpermu.test.hyperframe}
\title{Studentised Permutation Test}
\usage{
\method{studpermu.test}{hyperframe}(X, ...)
}
\description{
Studentised Permutation Test
}
\seealso{
\code{\link[spatstat]{studpermu.test}}
}
