% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{studpermu.test.pp3}
\alias{studpermu.test.pp3}
\title{Extends \code{\link[spatstat]{studpermu.test}} to pp3}
\usage{
studpermu.test.pp3(
  X,
  formula,
  summaryfunction = K3est,
  ...,
  nperm = 999,
  use.Tbar = FALSE,
  rinterval = NULL,
  minpoints = 20,
  rmax = NULL,
  nrval = 128,
  arguments.in.data = FALSE
)
}
\arguments{
\item{use.Tbar}{Logical value indicating choice of test statistic. If TRUE,
use the alternative test statistic, which is appropriate for summary
functions with roughly constant variance, such as K(r)/r or L(r). Defaults
to FALSE}

\item{rinterval}{Numeric of length 2. Experimental}
}
\description{
This function is not yet functional - many subfunctions must also be extended
to pp3.
}
\seealso{
\code{\link[spatstat]{studpermu.test}}

Other spatstat extensions: 
\code{\link{Tstat.pp3}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
