% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{studpermu.test.ppp}
\alias{studpermu.test.ppp}
\title{Studentised Permutation Test}
\usage{
\method{studpermu.test}{ppp}(
  X,
  formula,
  summaryfunction = Kest,
  ...,
  rinterval = NULL,
  nperm = 999,
  use.Tbar = FALSE,
  minpoints = 20,
  rsteps = 128,
  r = NULL,
  arguments.in.data = FALSE
)
}
\description{
Studentised Permutation Test
}
\seealso{
\code{\link[spatstat.core]{studpermu.test}}
}
