% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{summary_plotter_raw}
\alias{summary_plotter_raw}
\title{Plot summary functions}
\usage{
summary_plotter_raw(
  func = "K",
  observed_values,
  envelopes,
  ...,
  pattern.colors = c(`95.0\% AI` = "pink", Median = "black", Observed = "blue"),
  fill.colors = pattern.colors,
  unit = "nm",
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{func}{Summary function to plot}

\item{observed_values}{observed summary function value}

\item{envelopes}{theoretical values found by function such as `summary_functions().` Should be list}

\item{pattern.colors}{colors and names for envelope lines}

\item{fill.colors}{colors and names for envelope fill.  Match names to `pattern.colors`}

\item{K_cor}{edge correction(s) to be used for K function}

\item{G_cor}{edge correction(s) to be used for G function}

\item{F_cor}{edge correction(s) to be used for F function}

\item{GXGH_cor}{edge correction(s) to be used for GXGH function}

\item{GXHG_cor}{edge correction(s) to be used for GXHG function}
}
\description{
Plot the observed value with envelopes for expected values for summary function
}
