% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-extend.R
\name{superimpose.pp3}
\alias{superimpose.pp3}
\title{Extends \code{\link[spatstat]{superimpose}} to \code{\link[spatstat]{pp3}}.}
\usage{
\method{superimpose}{pp3}(..., W = NULL, check = F)
}
\description{
\code{superimpose.pp3}
}
\seealso{
\code{\link[spatstat]{superimpose}}
}
