% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{Tcalc}
\alias{Tcalc}
\title{Calculate T value}
\usage{
Tcalc(relabelings, func = "K", rmin = 0, rmax)
}
\description{
Takes input of relabeling object and calculates the T value for either K or G function
}
\details{
This uses the method from \emph{Baddeley et al.} to calculate the T value for an envelope for either
K  (\emph{K3est}) G \emph{G3est} function
}
\references{
Baddeley, A., Diggle, P. J., Hardegen, A., Lawrence, T., Milne, R. K., & Nair, G. (2014). On tests of spatial pattern based on simulation envelopes. Ecological Monographs, 84(3), 477–489. https://doi.org/10.1890/13-2042.1
}
