% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{tile_relabel}
\alias{tile_relabel}
\title{Tile and then relabel}
\usage{
tile_relabel(
  seed,
  pp3_full,
  funcs = c("K", "G", "F", "GXGH"),
  ...,
  host_formula,
  dopant_formula,
  x_tiles = 1,
  y_tiles = 1,
  z_tiles = 1,
  x_trim = 0,
  y_trim = 0,
  z_trim = 0,
  K_cor = "border",
  G_cor = "rs",
  F_cor = "rs",
  GXGH_cor = "rs",
  GXHG_cor = "rs"
)
}
\arguments{
\item{pp3_full}{input pp3 pattern for relabeling and tiling}

\item{host_formula, dopant_formula}{formula for}

\item{seeed}{input for `set.seed` function (for reproducibility)}

\item{summary}{functions to calculate on relabeling}
}
\description{
Tile and then relabel
}
