% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{tiler}
\alias{tiler}
\title{Tiler function}
\usage{
tiler(
  pattern,
  x_tiles = 1,
  y_tiles = 1,
  z_tiles = 1,
  x_trim = 0,
  y_trim = 0,
  z_trim = 0
)
}
\arguments{
\item{pattern}{input pp3 pattern}

\item{x_tiles, y_tiles, z_tiles}{number of tiles in each dimension (set equal to 1 for no change)}

\item{z_trim, y_trim, z_trim}{amount to trim off of each dimension.  Trimming will be applied equall to both sides}
}
\description{
This function takes an input pp3 pattern and tiles it in x, y, and z.  Then, it will trim it
according to the `trim` inputs
}
