% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{trueBox}
\alias{trueBox}
\title{Helper for \code{\link{makeCluster}} that determines a
\code{\link[spatstat]{pp3}} object true dimensions.}
\usage{
trueBox(pp3file)
}
\arguments{
\item{pp3file}{The \code{\link[spatstat]{pp3}} object to find the bounds of.}
}
\value{
A \code{\link[spatstat]{as.box3}} object containing the true volume
  dimensions.
}
\description{
RCP pattern generations, when loaded into R, lose their boundary information.
R interprets their boundary to be the extreme point locations in each
direction, when really they are usually nice integer values. This function
simply rounds the R boundary values to the true ones.
}
