% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{Tstat.pp3}
\alias{Tstat.pp3}
\title{Extends Tstat to pp3}
\usage{
Tstat.pp3(
  X,
  ...,
  rmax = NULL,
  correction = c("translation", "isotropic"),
  ratio = FALSE,
  verbose = TRUE
)
}
\description{
Extends Tstat to pp3
}
\seealso{
Other spatstat extensions: 
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
