% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{writeNode}
\alias{writeNode}
\title{Write a TAPSim Node File}
\usage{
writeNode(node, fp)
}
\arguments{
\item{node}{A \code{data.frame} with columns correponding to the
x, y, z, and id of each point in order}

\item{fp}{Character. The file to which to write the data.}
}
\value{
None.
}
\description{
\code{writeNode} writes a node file to the TAPSim format for creating a mesh
file for TAPSim.
\emph{Currently only ASCII is supported}
}
\seealso{
\url{
  http://www.uni-stuttgart.de/imw/mp/forschung/atom_probe_RD_center/howto.pdf
}

\url{https://d-nb.info/1138282715/34}
}
