% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_summarize.R
\name{average_relabelings}
\alias{average_relabelings}
\title{Average relabelings}
\usage{
average_relabelings(
  relabelings,
  envelope.value = 0.95,
  funcs = c("K", "G"),
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km",
  G2_cor = "km"
)
}
\arguments{
\item{relabelings}{output of  \code{\link[rTEM]{relabel_summarize}} function}

\item{envelope.value}{size of envelope to compute.  Should be decimal (e.g. 0.95 = 95\\%)}

\item{funcs}{vector of summary functions to calculate}

\item{K_cor}{edge correction(s) to be used for K function}

\item{G_cor}{edge correction(s) to be used for G function}

\item{F_cor}{edge correction(s) to be used for F function}

\item{GXGH_cor}{edge correction(s) to be used for GXGH function}

\item{GXHG_cor}{edge correction(s) to be used for GXHG function}
}
\value{
data frame with x value (distance), average y value, and y value envelopess
}
\description{
Function take all relabelings and return averages and envelope values
}
\details{
Take output of  \code{\link[rTEM]{relabel_summarize}} and create envelopes
}
