% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_process_functions.R
\name{bK3est}
\alias{bK3est}
\title{3D Border correction for K3est}
\usage{
bK3est(X, rmax = NULL, nrval = 128)
}
\arguments{
\item{X}{The point pattern for analysis. \code{\link[spatstat]{pp3}} object.}

\item{rmax}{See \code{\link[spatstat]{K3est}}. Maximum radius to be
calculated for \code{\link[spatstat]{K3est}}.}

\item{nrval}{See \code{\link[spatstat]{K3est}}. Number of radii that
\code{\link[spatstat]{K3est}} should be calculated at.}
}
\value{
Border corrected \code{\link[spatstat]{K3est}} data for object X.
}
\description{
Helper function for \code{\link{pK3est}}. This function is a hand written
extension of the border correction for 3D point patterns.
}
