% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{clustersim}
\alias{clustersim}
\title{Simulate clusters of marks on an RCP background point pattern.}
\usage{
clustersim(
  under,
  over,
  rcp_rad,
  pcp = 0.1,
  cr,
  rho1,
  rho2,
  rb = 0.1,
  pb = 0.1,
  tol = 0.005,
  s = 103,
  toplot = FALSE
)
}
\arguments{
\item{under}{The underlying RCP point pattern (UPP). A
\code{\link[spatstat.geom]{pp3}} object containing points at the centers of RCP
spheres.}

\item{over}{The overlying RCP pattern. A \code{\link[spatstat.geom]{pp3}} object
containing points at the centers of RCP spheres. Scaled equal to the
underlying point pattern.}

\item{rcp_rad}{The radius of the spheres within the underlying and overlying
RCP patterns.}

\item{pcp}{The fraction of type-A (clustering) points within the UPP. A
decimal value between 0 and 1.}

\item{cr}{Mean cluster radius. Any value larger than zero.}

\item{rho1}{Intra-cluster type-A point concentration. A decimal value between
0 and 1.}

\item{rho2}{Background type-A point concentration. A decimal value between 0
and the value of \code{pcp}.}

\item{rb}{Radius blur. A decimal value between 0 and 1.}

\item{pb}{Position blur. A decimal value between 0 and 1.}

\item{tol}{Tolerance value for \code{pcp}. The true fraction of type-A points
in the pattern will be within this tolerance of the value specified by the
\code{pcp} parameter. If not, function will return \code{NULL}.}

\item{s}{Random seed for the simulation.}

\item{toplot}{Show a 3D plot of the cluster points once generation is done?
\code{TRUE} or \code{FALSE} (the default).}
}
\value{
List of:
\itemize{
\item \code{clusters} - A \code{\link[spatstat.geom]{pp3}} object containing the final
locations of only type-A points within \code{all}.
\item \code{all} - A \code{\link[spatstat.geom]{pp3}} object containing the full marked
underlying point pattern with points marked as either type A (cluster type
point in a cluster), B (cluster type point not in a cluster) or C
(non-cluster type point).
\item \code{radii} - A vector containing the simulated radius of each cluster in
\code{all}.
\item \code{perc} - A single value containing the true fraction of type-A points in
\code{all}.
\item \code{centers} - A \code{\link[spatstat.geom]{pp3}} object containing the cluster
center locations within \code{all}.
}
}
\description{
\code{clustersim} simulates point clusters using two RCP point clouds. The first
point cloud is the "underlaying" pattern. This is the set of points that will
be used as actual cluster point locations in the resultant marked point
pattern. The second point cloud is the "overlaying" pattern. This is the set
of points that will be scaled to determine the positions of the cluster
centroids within the underlying point pattern.
}
\seealso{
Other simulation functions: 
\code{\link{lattice}()},
\code{\link{makecluster}()},
\code{\link{morph_lamellar}()},
\code{\link{morph_rods}()},
\code{\link{nmers}()},
\code{\link{read.rcp}()},
\code{\link{scaleRCP}()},
\code{\link{stitch.size}()}
}
\concept{simulation functions}
