% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{createSpat}
\alias{createSpat}
\title{Create a pp3 from a POS or ATO}
\usage{
createSpat(pos, win = NULL, marks = NULL)
}
\arguments{
\item{pos}{A POS or ATO data frame.}

\item{win}{The domain of the data.}
}
\value{
A \code{\link[spatstat.geom]{pp3}} with the x,y,z positions of the hits in
the supplied POS or ATO.
}
\description{
\code{createSpat} creates a \code{\link[spatstat.geom]{pp3}} from a POS or ATO
data frame.
}
\seealso{
\code{\link{readPOS}}, \code{\link{readATO}},
\code{\link[spatstat.geom]{pp3}}

Other APT data transformation functions: 
\code{\link{createDet}()},
\code{\link{createSpec}()}
}
\concept{APT data transformation functions}
