% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{createSpec}
\alias{createSpec}
\title{Create a \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} from a POS
or ATO}
\usage{
createSpec(pos, res = 0.05, snip = NULL)
}
\arguments{
\item{pos}{A POS or ATO data frame.}

\item{res}{Numeric. The bin width of the mass spectrum.}

\item{snip}{Ignored.
\code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} resolution.}
}
\value{
A \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} from the
  \code{mass} field of the POS or ATO, with the resolution set by \code{res}.
}
\description{
\code{createSpec} generates a
\code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} object with a
specified resolution from an ATO or POS data frame (like that created by
\code{\link{readPOS}}).
}
\details{
The input POS or ATO is binned by mass values; the resolution parameter sets
the width of the mass bins used in \code{\link[graphics]{hist}} to create the
input to the \code{\link[MALDIquant]{createMassSpectrum}} call, and also acts
as a tolerance around the spectrum minimum and maximum mass. The minimum of
the mass value is zero.
}
\seealso{
\code{\link{readPOS}}, \code{\link{readATO}},
  \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}
}
