% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-file.R
\name{createSpec}
\alias{createSpec}
\title{Create a \code{\link[MALDIquant]{MassSpectrum}} from a POS or ATO
data frame.}
\usage{
createSpec(pos, res = 0.001, snip = NULL)
}
\arguments{
\item{pos}{A POS or ATO data frame.}

\item{res}{The desired \code{\link[MALDIquant]{MassSpectrum}} resolution.}
}
\value{
A \code{\link[MALDIquant]{MassSpectrum}} from the \code{mass} field
  of the POS or ATO, with the resolution set by \code{res}.
}
\description{
\code{createSpec} generates a \code{\link[MALDIquant]{MassSpectrum}} object
with a specified resolution from an ATO or POS data frame (like that created
by \code{\link(readPOS)}).
}
\seealso{
\code{\link{readPOS}}, \code{\link{readATO}},
  \code{\link[MALDIquant]{MassSpectrum}}
}
