% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-file.R
\name{createSpec}
\alias{createSpec}
\title{Create a \code{\link[MALDIquant]{MassSpectrum}} object from a POS or ATO
data frame.}
\usage{
createSpec(pos, res = 0.001)
}
\description{
Create a \code{\link[MALDIquant]{MassSpectrum}} object from a POS or ATO
data frame.
}
\seealso{
\code{\link{readPOS}}, \code{\link{readATO}}
}
