% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{drawRange}
\alias{drawRange}
\title{Color ranges on a MassSpectrum}
\usage{
drawRange(ms, start, end, col = "red")
}
\description{
\code{drawRange} draws a colored range fon a plot of a
\code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}
}
\seealso{
\code{\link{drawRNG}}
}
