% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{drawRange}
\alias{drawRange}
\title{Color ranges on a MassSpectrum}
\usage{
drawRange(ms, start, end, col = "red", miny = 0)
}
\arguments{
\item{ms}{The \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} or
\code{data.frame} (with columns named \code{mass} and \code{intensity}) on which the ranges
are to be drawn}

\item{start}{numeric. The start of the range(s).}

\item{end}{numeric. The end of the range(s).}

\item{col}{The color of the range(s).}
}
\value{
\code{NULL}
}
\description{
\code{drawRange} draws a colored range on a plot
}
\seealso{
\code{\link{drawRNG}}
}
