% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{drawRNG}
\alias{drawRNG}
\title{Color ranges from an RNG on a MassSpectrum}
\usage{
drawRNG(ms, rng)
}
\arguments{
\item{ms}{The \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} or
\code{data.frame} (with columns named \code{mass} and \code{intensity}) on which the ranges
are to be drawn}

\item{rng}{The RRNG \code{data.frame} used to color the ranged peaks in the
mass spectrum}
}
\value{
\code{NULL}
}
\description{
\code{drawRNG} draws colored ranges from an RNG (as generated by
\code{\link{readRNG}}) on a plot
}
\seealso{
\code{\link{drawRange}}
}
