% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_functions.R
\name{envPlot}
\alias{envPlot}
\title{Plot envelopes of K3est test}
\usage{
envPlot(tests, percentiles = c(0.999, 0.99, 0.97), ylim = c(-3, 3),
  xlim = c(0, ceiling(max(tests[, 1]))), leg = TRUE, ...)
}
\arguments{
\item{tests}{The return file from \code{p(K/G/F)3est} or the first, [[1]],
entry in the list returned by \code{p(K/G/F)3est} with \code{anom = TRUE}.}

\item{percentiles}{Numerical vector of percentiles that you want to see the
envelopes for. Each between 0 and 1.}

\item{ylim}{Numerical vector containing the min and max values for the y axis
on the plot.}

\item{xlim}{Numerical vector containing the min and max values for the x axis
on the plot.}

\item{leg}{True or falsel whether to show the automatically generated legend.}

\item{...}{Arguments to be passed into \code{plot()}.}

\item{colors}{List of color names to make the envelopes.}
}
\value{
Nothing.
}
\description{
Plot the results of envelope calculations from the \code{\link{pK3est}},
\code{\link{pG3est}}, or \code{\link{pF3est}} functions, with the ability to
choose the percentiles for plotting.
}
