% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-detection.R
\name{gema.pp3}
\alias{gema.pp3}
\title{Identify Clusters in a Marked Point Pattern Using GEMA}
\usage{
\method{gema}{pp3}(X, threshold = 0.5)
}
\arguments{
\item{X}{The marked point pattern (object of class
\code{\link[spatstat]{pp3}}) in which to identify clusters.}

\item{threshold}{The probability threshold at which to assign a point to a
cluster or to the background.}
}
\value{
A marked point pattern (object of class \code{\link[spatstat]{pp3}})
with marks corresponding to the identified background and each identified
cluster.
}
\description{
Identify Clusters in a Marked Point Pattern Using GEMA
}
\details{
A \code{data.frame} is returned as an attribute (\code{attr("prob")}) that contains the
probabilities of assignment to the background and each cluster.
}
\references{
Zelenty, J. \emph{et al.},
"Detecting Clusters in Atom Probe Data with Gaussian Mixture Models",
\emph{Microscopy and Microanalysis}, \strong{23} (2), 269-278 (2017):
\url{https://doi.org/10.1017/S1431927617000320}
}
\seealso{
Other cluster identification functions: 
\code{\link{gema.gema}()},
\code{\link{msa}()}
}
\concept{cluster identification functions}
