% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_spec.R
\name{ionFormula}
\alias{ionFormula}
\title{Helper Function for \code{\link{fitIonInit}}}
\usage{
ionFormula(ions, charge = 1, threshold = 10, peak = "gaussian")
}
\arguments{
\item{ions}{Character. Molecular formula formatted in a way that is
acceptable to \code{\link[enviPat]{isopattern}}}

\item{charge}{Numeric. Charge of ion.}

\item{threshold}{Numeric. Percentage threshold of isotope.}

\item{peak}{Character. One of "gaussian" or "emg". See Details.}
}
\value{
A character string that concatinates all the formulae for the ions
  and their respetive isotopes, along with a constant noise term.
}
\description{
Creates the formulae that get passed to \code{\link[stats]{nls}}
}
\details{
The base peak shape of the formula is either a gaussian
  (peak = "gaussian") or an exponentially modified gaussian (peak = "emg").
}
\seealso{
\code{\link[enviPat]{fitIonInit}}, \code{\link[enviPat]{isopattern}}

Other spectral fitting functions: 
\code{\link{fitIonInit}()}
}
\concept{spectral fitting functions}
