% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_vis.R
\name{isoLabel}
\alias{isoLabel}
\title{Label a prettyPlot with isotope patterns}
\usage{
isoLabel(ms, pat, rng = NULL, col = NULL, halfWindowSize = 0.5)
}
\arguments{
\item{ms}{A \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}}. The
(log transformed) mass spectrum to be marked.}

\item{pat}{A list of isotopic patterns as generated by
\code{\link[enviPat]{isopattern}}}

\item{rng}{Currently unused.}

\item{col}{A vector of colors for the isotopes. If \code{NULL}, the standard R
sequence of colors will be used.}

\item{halfWindowSize}{A numeric. The half-width of the window for determining
the height of the main isotope peak.}
}
\value{
\code{NULL}
}
\description{
\code{isoLabel} adds a set of lines and labels of isotopes generated by
\code{\link[enviPat]{isopattern}} to a \code{\link{prettyPlot}}. It currently
assumes (like \code{prettyPlot}) that the spectrum is log10 transformed. This will
be extended to arbitrary spectra in the future.
}
\seealso{
\code{\link[enviPat]{isopattern}}
}
