% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{kseries2}
\alias{kseries2}
\title{Helper function to allow for parallelization of metric extraction from
cluster simulations}
\usage{
kseries2(j, p, tot, maxr, nr, toSub)
}
\arguments{
\item{j}{The index of the RCP pattern to run (note that this might only be
able to be applied on the mines HPC under Galen's account right now as
written)}

\item{p}{The total number of RCP patterns available.}

\item{tot}{A list of cluster properties to test. This should be a list of
vectors containing c(r, den, rb, gbp).}

\item{maxr}{Maximum r to calculate K function to.}

\item{nr}{Number of r values to evaluate k function at.}

\item{toSub}{A vector of values to substitute for the \code{\link{anomK3est}}
function.}
}
\value{
Matrix containing 5 metrics and the seed value for each parameter
  combination given in the \code{tot} list.
}
\description{
This function takes a single RCP pattern and creates and tests the K function
metrics on different cluster properties passed into the function. Used to
parallelize large sets of cluster property runs.
}
