% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{makecluster}
\alias{makecluster}
\title{Simulate point clustering in an RCP matrix}
\usage{
makecluster(under, over, radius1, radius2, type = "ppc", ppc = NULL,
  cr = NULL, slow = FALSE, fast = FALSE, superfast = FALSE,
  d = NULL, pic = 1, pcp = 0.06, toPlot = FALSE,
  showOverPts = FALSE)
}
\arguments{
\item{under}{The underlaying RCP pattern. A \code{\link[spatstat]{pp3}}
object containing the correctly scaled RCP pattern point locations. Should
have used scaleRCP prior to putting hte object into this argument.}

\item{over}{The overlaying RCP pattern. A \code{\link[spatstat]{pp3}} object
containing the RCP pattern point locations. Scaled equal to the underlaying
pattern.}

\item{radius1}{The small radius of the underlaying RCP pattern. Can be found
in the system output file, or as whatever the pattern is scaled to.}

\item{radius2}{The small radius of the overlaying RCP pattern. Can be found
in the system output file, or as whatever the pattern is scaled to.}

\item{type}{How to create the clusters. Either "ppc", "cr", or "dist". See
below for more information on each.}

\item{ppc}{Number of points per cluster if type = "ppc", otherwise NULL.}

\item{cr}{Cluster radius if type = "cr", otherwise NULL.}

\item{fast}{Whether or not to use the "fast" cluster radius method if type =
"cr".}

\item{d}{Distance between clusters if type = "dist", otherwise NULL.}

\item{pic}{Percent In Clusters. Percent of the points marked as cluster type
that should actually be contained within the clusters. Number between 0 and
1 For example, if pic = 0.5, 50 percent of the cluster type points will be
in clusters and 50 percent will be randomly spread through the point cloud.}

\item{pcp}{Percent Cluster Points. Percent of the points in the underlaying
RCP pattern that should be marked as cluster type points. Number between 0
and 1.}

\item{toPlot}{Show a 3D plot of the cluster points once generation is done?
TRUE or FALSE.}

\item{showOverPts}{If toPlot = TRUE, show all of the points in the RCP
pattern, with clustered points marked in red, or just show the clustered
points? TRUE or FALSE.}
}
\value{
Returns are different based on \code{type}.
  \subsection{\code{type} = "ppc" or "dist"}{List of:
  [[1]] \code{\link[spatstat]{pp3}} object containing the cluster marked
  point locations. [[2]] \code{\link[spatstat]{pp3}} object containing the
  overlaying RCP pattern after scaling. [[3]] Numeric vector containing: [1]
  points per cluster 1 [2] number of points with points per cluster 1 [3]
  points per cluster 2 [4] number of points with points per cluster 2. }
  \subsection{\code{type} = "cr", fast = FALSE}{List of: [[1]]
  \code{\link[spatstat]{pp3}} object containing the cluster marked point
  locations. [[2]] \code{\link[spatstat]{pp3}} object containing the
  overlaying RCP pattern after scaling. [[3]] Numeric vector containing: [1]
  number of clusters [2] number of points in those clusters.}
  \subsection{\code{type} = "cr", fast = TRUE}{List of: [[1]]
  \code{\link[spatstat]{pp3}} object containing the cluster marked point
  locations. [[2]] \code{\link[spatstat]{pp3}} object containing the
  overlaying RCP pattern after scaling. [[3]] factor containing the number of
  points in each cluster.}
}
\description{
The \code{makecluster} function simulates point clusters using two RCP point
clouds. The first point cloud is the "underlaying" pattern. This is the set
of points that will be used as actual cluster point locations in the final
product. The second point cloud is the "overlaying" pattern. This is the set
of points that will determine the positions of the clusters within the
underlaying pattern.
}
\section{Cluster Creation Methods}{
 What the different arguments for
  \code{type} argument mean.
  \subsection{Points Per Cluster - "ppc"}{
  Specify the number of points that should be in each cluster. Needs to be
  paired with the \code{ppc} argument The program will scale the overlaying
  pattern so that there are approximately the correct number of clusters
  placed through the underlaying patten, mark the n closest points to these
  cluster centers as cluster points, where n is the value input to the ppc
  argument. The cluster centers are defined by the point locations of the
  scaled overlaying RCP pattern. At this point, there should be approximately
  N x \code{pcp} x \code{pip} points containined in clusters (N = total
  points in the underlaying RCP pattern). The method cleans up by marking or
  removing marks around the centers as necesary so that there are exactly N x
  pcp x pip points in clusters. The rest of the N x \code{pcp} x
  (1-\code{pip}) points are then placed randomly through the remaining
  non-cluster-marked points.}
  \subsection{Cluster Radius - "cr"}{
  Specify the radius of the clusters. Needs to be paried with the \code{cr}
  argument. The overlaying RCP pattern will be scaled so that there are
  approximately the correct number of clusters placed through the underlaying
  pattern. All points in the underlaying pattern within distance \code{cr} of
  the cluster centers are marked as cluster points. The cluster centers are
  defined by the point locations of the scaled overlaying RCP pattern. At
  this point, there should be approximately N x \code{pcp} x \code{pip}
  points containined in clusters (N = total points in the underlaying RCP
  pattern). The method cleans up by marking or removing marks around the
  centers as necesary so that there are exactly N x pcp x pip points in
  clusters. The rest of the N x \code{pcp} x (1-\code{pip}) points are then
  placed randomly through the remaining non-cluster-marked points.

  If fast = FALSE: Uses the original, hand written function
  \code{\link{nncrossR}} to calculate points within a given radius. This
  method is slow, but very accurate. If fast = TRUE: Uses the spatstat
  function \code{\link[spatstat]{closepairs.pp3}} to find points within a
  given radius. Faster, but slightly less accurate.}
  \subsection{Distance - "dist"}{
  Specify the distance between clusters. Needs to be paired with the \code{d}
  argument. The overlaying RCP pattern will be scaled according to the
  \code{d} argumnet so that the small radius of the overlaying RCP pattern is
  \code{d}/2. Selects the closest n points around each cluster center so that
  there are N x \code{pcp} x \code{pip} points contained in clusters. The
  rest of the N x \code{pcp} x (1-\code{pip}) points are then placed randomly
  through the remaining non-cluster-marked points.}
}

