% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{marktable.pp3}
\alias{marktable.pp3}
\title{Tabulate Marks in Neighbourhood of Every Point in a Point Pattern}
\usage{
\method{marktable}{pp3}(X, R, N, exclude = TRUE, collapse = FALSE)
}
\description{
Visit each point in a point pattern, find the neighbouring
  points, and compile a frequency table of the marks of these neighbour
  points.
}
