% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{marktable}
\alias{marktable}
\title{Tabulate Marks in Neighbourhood of Every Point in a Point Pattern}
\usage{
marktable(X, ...)
}
\description{
This is an S3 generic that extends the use of
  \code{\link[spatstat]{marktable}} beyond \code{ppp} objects.
}
\seealso{
\code{\link[spatstat]{marktable}}, \code{\link{marktable.ppp}},
  \code{\link{marktable.pp3}}
}
