% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{marktable}
\alias{marktable}
\title{Tabulate Marks in Neighbourhood of Every Point in a Point Pattern}
\usage{
marktable(X, ...)
}
\arguments{
\item{X}{A marked point pattern. An object of class "ppp".}

\item{R}{Neighbourhood radius. Incompatible with \code{N}.}

\item{N}{Number of neighbours of each point. Incompatible with \code{R}.}

\item{exclude}{Logical. If \code{exclude=TRUE}, the neighbours of a point do
not include the point itself. If \code{exclude=FALSE}, a point belongs to
its own neighbourhood.}

\item{collapse}{Logical. If \code{collapse=FALSE} (the default) the results
for each point are returned as separate rows of a table. If
\code{collapse=TRUE}, the results are aggregated according to the type of
point.}
}
\description{
This is an S3 generic that extends the use of
\code{\link[spatstat]{marktable}} beyond "ppp" objects.
}
\seealso{
\code{\link[spatstat]{marktable.ppp}}, \code{\link{marktable.pp3}}

Other spatstat extensions: 
\code{\link{G3cross}()},
\code{\link{G3multi}()},
\code{\link{K3scaled}()},
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
