% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_max-sep-algorithm.R
\name{msa}
\alias{msa}
\title{Performs the maximum separation algorithm on a marked point pattern.}
\usage{
msa(X, dmax, Nmin, denv, der, clust.mark = c("A"))
}
\arguments{
\item{X}{A marked \code{\link[spatstat]{pp3}} object which MSA will be
performed on.}

\item{dmax}{Maximum separation distance; The maximum distance two points can
be separated by and still be considered part of the same cluster.}

\item{Nmin}{Minimum cluster points; The minimum number of points needed to
classify a grouping of points as a cluster.}

\item{denv}{Envelope distance; Any points within this distance of a point
residing in a cluster will be included in the cluster (this controls the
addition of background points to a cluster).}

\item{der}{Erosion distance; Any points within this distance of a background
matrix point (after enveloping) will be removed from the cluster.}

\item{clust.mark}{Vector containing the names of the marks in \code{X} that
should be considered as cluster type points. All points with marks not
included in this vector will be considered background points.}
}
\value{
A list of [[1]] Vector containing estimated radius of each cluster
  found, [[2]] Vector containing estimated intra-cluster concentration of
  cluster type points in each cluster found, [[3]] Estaimted background
  concentration of cluster type points within the background (the entire
  domain minus the clusters), [[4]] List of indices from the original pattern
  of cluster type points that reside in clusters, [[5]] List of indices from
  the original pattern of background type points that reside in clusters.
}
\description{
Main argument is a marked \code{\link[spatstat]{pp3}} object. Marks can be of
multiple types, but will be reduced to only two (the cluster species and
non-cluster species) as part of the algorithm.
}
