% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim.R
\name{nmers}
\alias{nmers}
\title{Select n-mer Clusters}
\usage{
nmers(sam, parent, n = 2)
}
\description{
Creates an n-mer point pattern by selecting the n+1 nearest neighbors of a
sampled point pattern from its parent distribution.
}
