% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_NNDE.R
\name{nncrossden.pp3}
\alias{nncrossden.pp3}
\title{Similar to \code{\link{nndensity.pp3}}, but specifically for marked patterns
with a global intensity inhomogeneity.}
\usage{
nncrossden.pp3(
  X,
  Y,
  k,
  nx,
  ny,
  nz,
  at.points = FALSE,
  nsplit = 1000,
  cores = 8,
  os = "linux"
)
}
\arguments{
\item{X}{The point pattern to estimate the intensity of.}

\item{Y}{The full point pattern to estimate global intensity.}

\item{k}{Vector containing the nearest-neighbor #s that the estimate should
be calculated for.}

\item{nx, ny, nz}{If estimating on a grid, the number of grid points in x, y,
and z.}

\item{at.points}{\code{TRUE} or \code{FALSE}. Whether or not to estimate
intensity at points in pattern. If \code{TRUE}, nx, ny, and nz are not
used.}

\item{nsplit}{In this function, large \code{pp3} objects are split into
multiple smaller data sets so that the memory is not overloaded while doing
computations. This parameter is the number of points per split set.}

\item{cores}{Number of cores to use for parallelization. Set to one for
serial calculation.}

\item{os}{Either 'windows', 'mac', or 'linux'. Changes the parallelization
method used.}
}
\value{
List containing: [\link{1}] A data frame of the intensity estimates for
each nearest neighbor value. [\link{2}] The coordinates of the estimates. [\link{3}]
The coordinates of the original points from the data set. [\link{4}] A vector
containing all k values tested.
}
\description{
Calculates the 3D nearest-neighbor intensity estimate of a point process at
either a grid of points or at the point locations in the data set. Requires
two point patterns: One (\code{X}) that contains just points of the type you
want to estimate the intesnity of. A second (\code{Y}) that contains all
points from the sample (note that this includes the points from pattern
\code{X}) for removing global intensity. Note that this function runs faster
on linux and mac machines than windows, but will work on both.
}
