% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_recon.R
\name{paraRec}
\alias{paraRec}
\title{Reconstruction using a parabolic tip shape assumption.}
\usage{
paraRec(ato, a, d = 9e+08, icf = 1)
}
\description{
\code{paraRec} creates a reconstruction based upon a parabolic tip shape
assumption and parabolic flight paths.
}
