% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-vis.R
\name{prettyPlot}
\alias{prettyPlot}
\title{Plot a pretty \code{\link[MALDIquant]{MassSpectrum}}.}
\usage{
prettyPlot(ms, rng = NA, xlim = NULL, ylim = NULL, lwd = 1,
  main = "Mass Spectrum", hold.par = F)
}
\description{
\code{prettyPlot}
}
