% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{quadratcount.pp3}
\alias{quadratcount.pp3}
\title{Extension of \code{\link[spatstat]{quadratcount}} to \code{\link[spatstat]{pp3}} objects.}
\usage{
\method{quadratcount}{pp3}(X, nx = 5, ny = 5, nz = 5)
}
\arguments{
\item{X}{The \code{\link[spatstat]{pp3}} object to split up.}

\item{nx, ny, nz}{Number of ractangular quadrats in the x, y, and z directions.}
}
\value{
A \code{data.frame} object containing the number of counts in each quadrat.
}
\description{
Divides volume into quadrats and counts the number of points in each quadrat.
}
