% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{quadrats.pp3}
\alias{quadrats.pp3}
\title{Divide a 3D Point Pattern into Sub-Volumes}
\usage{
quadrats.pp3(X, nx = 5, ny = nx, nz = ny, box.dims = NULL)
}
\arguments{
\item{X}{The \code{\link[spatstat]{pp3}} object to split up.}

\item{nx, ny, nz}{Number of rectangular quadrats in the x, y, and z directions,
if you wish to split up your point patthern by number of boxes.}

\item{box.dims}{Vector containing the dimensions of the subsetted 3D boxxes,
if you wish to define the individual boxx size. Use either \code{nx}, \code{ny},
\code{nz} or \code{box.dims}, but not both. See Details.}
}
\value{
A list containing the split up "pp3" objects.
}
\description{
Divides volume into quadrats and returns them.
}
\details{
If \code{box.dims} is not commensurate with the size of the object, the
quadrats are justified toward the smallest boundary in each dimension
(\emph{i.e.} (\code{c(min(x), min(y), min(z))})).
}
\seealso{
\code{\link[spatstat]{quadrats}}

Other spatstat extensions: 
\code{\link{G3cross}()},
\code{\link{G3multi}()},
\code{\link{K3scaled}()},
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
