% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{randomInsert}
\alias{randomInsert}
\title{Helper for \code{\link{makecluster}} to insert random cluster points}
\usage{
randomInsert(cluster.indices, n, N, s, points.avoid = cluster.indices)
}
\arguments{
\item{cluster.indices}{A vector containing the indices of the current cluster
points}

\item{n}{The number of points that need to be placed randomly}

\item{N}{The number of points in the entire underlaying pattern.}

\item{s}{The seed for the random inserted points.}

\item{points.avoid}{A vector containing the indices of the points you don't
want the function to insert points to.}
}
\value{
New indices vector containing new cluster points randomly placed.
}
\description{
When \code{pip} argument of \code{\link{makecluster}} is not equal to 1, or
\code{superfast == TRUE}, there may be random points that need to be marked
as cluster type placed within the underlaying pattern. This function does
just that.
}
