% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{randomInsert}
\alias{randomInsert}
\title{Helper for \code{\link{makecluster}} to insert random cluster points}
\usage{
randomInsert(cluster.Indices, n, N)
}
\arguments{
\item{cluster.Indices}{A vector containing the indices of the current cluster
points}

\item{n}{The number of points that need to be placed randomly}

\item{N}{the number of points in the entire underlaying pattern.}
}
\value{
New indices vector containing new cluster points randomly placed.
}
\description{
When \code{pip} argument of \code{\link{makecluster}} is not equal to 1,
there are random points that need to be marked as cluster type placed within
the underlaying pattern. This function does just that.
}
