% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_cluster-sim.R
\name{randomTakeAway}
\alias{randomTakeAway}
\title{Helper for \code{\link{makecluster}} to take away random cluster points}
\usage{
randomTakeAway(cluster.indices, n, N, s)
}
\arguments{
\item{cluster.indices}{A vector containing the indices of the current cluster
points}

\item{n}{The number of points that need to be removed randomly}

\item{N}{The number of points in the entire underlaying pattern.}

\item{s}{The random seed for the random selection of points to take away.}
}
\value{
New indices vector containing new cluster points randomly placed.
}
\description{
When \code{superfast} argument of \code{\link{makecluster}} is \code{TRUE},
there may be random points that need to be removed from the existing clusters
of the the underlaying pattern. This function does just that.
}
