% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{rangeCount}
\alias{rangeCount}
\title{Count the hits within a mass range}
\usage{
rangeCount(pos, start, end)
}
\arguments{
\item{pos}{A data.frame. The pos to be ranged}

\item{start}{The start of the mass range}

\item{end}{The end of the mass range}
}
\value{
The number of hits falling within the range.
}
\description{
To get the hit information within a mass range, use \code{\link{rangePOS}}
}
\seealso{
\code{\link{rngCount}}, \code{\link{rangePOS}}
}
