% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{rangePOS}
\alias{rangePOS}
\title{Extract hits within a given mass range.}
\usage{
rangePOS(pos, start, end)
}
\arguments{
\item{pos}{A data.frame. The pos to be ranged}

\item{start}{The start of the mass range}

\item{end}{The end of the mass range}
}
\value{
A data.frame of the same structure as \code{pos} containing only hits
in the provided range

A data.frame of the same structure as \code{pos} containing only hits
in the provided range
}
\description{
rangePOS extracts the rows of a \code{POS} or \code{ATO} object whose mass
is within the provided range.
}
