% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{read.rcp}
\alias{read.rcp}
\title{Read in an RCP simulation.}
\usage{
read.rcp(fpath_config, fpath_sys, scaleUp, newRadius = 0.5)
}
\arguments{
\item{fpath_config}{The file path to the RCP FinalConfig file.}

\item{fpath_sys}{The file path to the associated RCP system file.}

\item{newRadius}{If \code{scaleUP = TRUE}, this is the new radius that the
RCP particles will be scaled to. Default is 0.5.}

\item{scaleUP}{Boolean. If \code{TRUE}, scales RCP so particles have new
radius. If \code{FALSE}, RCP stays as generated.}
}
\value{
A \code{\link[spatstat]{pp3}} object of the RCP pattern.
}
\description{
Function to read in an RCP code output file into R as a pp3 object.
}
