% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{readATO}
\alias{readATO}
\title{Read an ATO File}
\usage{
readATO(filepath)
}
\arguments{
\item{filepath}{Character. The file path to the ATO file.}
}
\value{
A data.frame with columns corresponding to the ATO structure
  definition.
}
\description{
\code{readATO} reads an ATO file (from IVAS) into a data frame.
}
\references{
Larson \emph{et al.},
  \emph{Local Electrode Atom Probe Tomography: A User's Guide} (2009)
}
\seealso{
Other APT import functions: 
\code{\link{readPOS}()},
\code{\link{readRRNG}()}
}
\concept{APT import functions}
