% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{readATO}
\alias{readATO}
\title{Read an ATO file.}
\usage{
readATO(filepath)
}
\arguments{
\item{filepath}{A string. The file path to the ATO file.}
}
\value{
A dataframe.
}
\description{
\code{readATO} reads an ATO file (from IVAS) into a data frame.
}
\references{
Local Electrode Atom Probe Tomography: A User's Guide
}
\seealso{
\code{\link{readPOS}}
}
