% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-file.R
\name{readForm}
\alias{readForm}
\title{Create a list of mass spec formulae from a table (class \code{cform}).}
\usage{
readForm(filepath)
}
\description{
Create a list of mass spec formulae from a table (class \code{cform}).
}
