% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-file.R
\name{readPOS}
\alias{readPOS}
\title{Read a POS file.}
\usage{
readPOS(filepath)
}
\arguments{
\item{filepath}{A string. The file path to the POS file.}
}
\description{
\code{readPOS} reads a POS file (from IVAS) into a data frame.
}
\seealso{
\code{\link{readPOS}}
}
