% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{readPOS}
\alias{readPOS}
\title{Read a POS File}
\usage{
readPOS(filepath)
}
\arguments{
\item{filepath}{Character. The file path to the POS file.}
}
\value{
A \code{data.frame} with columns corresponding to the x, y, and z
positions of the reconstruction and the mass-to-charge ratio.
}
\description{
\code{readPOS} reads a POS file (from IVAS) into a \code{data.frame}.
}
\references{
Larson \emph{et al.},
\emph{Local Electrode Atom Probe Tomography: A User's Guide} (2009)
}
\seealso{
Other APT import functions: 
\code{\link{readATO}()},
\code{\link{readRRNG}()}
}
\concept{APT import functions}
