% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{readResult}
\alias{readResult}
\title{Read a TAPSim results file}
\usage{
readResult(fp, type = "ASCII")
}
\arguments{
\item{fp}{The filepath}

\item{type}{The file type to be read. Currently, only ASCII is implemented.}
}
\value{
A data.frame
}
\description{
Given a filepath, read in a TAPSim results file. Currently only implemented
for ASCII results files.
}
\seealso{
\url{
  http://www.uni-stuttgart.de/imw/mp/forschung/atom_probe_RD_center/howto.pdf
}

\url{https://d-nb.info/1138282715/34}
}
