% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-file.R
\name{readRRNG}
\alias{readRRNG}
\title{Read an RRNG file (from IVAS) into a dataframe}
\usage{
readRRNG(filepath)
}
\arguments{
\item{filepath}{A string. A filepath to the RRNG file.}
}
\value{
A dataframe.
}
\description{
\code{readRRNG} reads an RRNG file from IVAS into an RNG object for ranging
a \code{\link[MALDIquant:MassSpectrum-class]{MassSpectrum}} generated by
\code{\link{createSpec}}
}
