% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{resultToDet}
\alias{resultToDet}
\title{Convert TAPSim Result to Detector "ppp"}
\usage{
resultToDet(res, clip.radius = NULL)
}
\arguments{
\item{res}{The results \code{data.frame} as returned by \code{\link{readResult}}}

\item{clip.radius}{Numeric. The detector radius at which to clip the points.
If \code{NULL} (the default), no clipping is applied.}
}
\value{
A \code{\link[spatstat.geom]{ppp}} containing the detector positions
marked by their id.
}
\description{
\code{resultToDet} converts a TAPSim result \code{data.frame} (as returned by
\code{\link{readResult}}) into a \code{\link[spatstat.geom]{ppp}} using
\code{\link[=createDet]{createDet()}}.
}
\seealso{
\code{\link{createDet}}

Other TAPSim functions: 
\code{\link{readResult}()},
\code{\link{resultToPOS}()}
}
\concept{TAPSim functions}
