% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{resultToDet}
\alias{resultToDet}
\title{Convert TAPSim Result to Detector \code{\link[spatstat]{ppp}}}
\usage{
resultToDet(res, clip.radius = NULL)
}
\arguments{
\item{res}{data.frame. The results \code{data.frame} as returned by
\code{\link{readResult}}}

\item{clip.radius}{Numeric. The detector radius at which to clip the points.}
}
\value{
A \code{\link[spatstat]{ppp}} containing the detector positions
  marked by their id.
}
\description{
Convert TAPSim Result to Detector \code{\link[spatstat]{ppp}}
}
\seealso{
\code{\link{createDet}}

Other TAPSim functions: 
\code{\link{readResult}()},
\code{\link{resultToPOS}()}
}
\concept{TAPSim functions}
