% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_tapsim.R
\name{resultToPOS}
\alias{resultToPOS}
\title{Convert TAPSim Results to POS}
\usage{
resultToPOS(res, clip.radius = NULL)
}
\arguments{
\item{res}{The results \code{data.frame} as returned by \code{\link{readResult}}}

\item{clip.radius}{Numeric. The detector radius at which to clip the points}
}
\value{
A \code{data.frame} containing the (x,y,z) position and "mass" (id)
of the point in the stame structure of a POS as created by
\code{\link{readPOS}}.
}
\description{
\code{resultToDet} converts a TAPSim result \code{data.frame} (as returned by
\code{\link{readResult}}) into a \code{\link[spatstat.geom]{pp3}} using
\code{\link[=readPOS]{readPOS()}}.
}
\seealso{
\code{\link{readPOS}}

Other TAPSim functions: 
\code{\link{readResult}()},
\code{\link{resultToDet}()}
}
\concept{TAPSim functions}
