% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{rjitter.pp3}
\alias{rjitter.pp3}
\title{Random Perturbation of a Point Pattern}
\usage{
rjitter.pp3(X, domain = box3())
}
\description{
Applies independent random displacements to each point in a point pattern.
Extends \code{\link[spatstat]{rjitter}} to \code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[spatstat]{rjitter}}, \code{\link{rjitter.ppp}}
}
