% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{rjitter.pp3}
\alias{rjitter.pp3}
\title{Random Perturbation of a Point Pattern}
\usage{
\method{rjitter}{pp3}(X, domain = box3())
}
\description{
Applies independent random displacements to each point in a point pattern.
Extends \code{\link[spatstat]{rjitter}} to \code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[spatstat]{rjitter}}, \code{\link{rjitter.ppp}}

Other spatstat extensions: 
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
