% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{rjitter.ppp}
\alias{rjitter.ppp}
\title{Random Perturbation of a Point Pattern}
\usage{
\method{rjitter}{ppp}(X, radius, retry = TRUE, giveup = 10000, ..., nsim = 1, drop = TRUE)
}
\description{
Random Perturbation of a Point Pattern
}
\seealso{
\code{\link[spatstat]{rjitter}}, \code{\link{rjitter.pp3}}

Other spatstat extensions: 
\code{\link{Tstat.pp3}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
