% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spec.R
\name{rngCount}
\alias{rngCount}
\title{Count the number of hits for each entry within a \code{RNG} object}
\usage{
rngCount(pos, rng)
}
\description{
Count the number of hits for each entry within a \code{RNG} object
}
\seealso{
\code{\link{readRRNG}}, \code{\link{rangeCount}}
}
