% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_spec.R
\name{rngCount}
\alias{rngCount}
\title{Count Hits in RRNG}
\usage{
rngCount(pos, rng, simplify = FALSE)
}
\arguments{
\item{pos}{The pos to range}

\item{rng}{The ranges}
}
\value{
A data.frame containing the name of each range, the number of counts
and fraction of the total ranged counts.
}
\description{
\code{rngCount} counts the number of hits for each entry within a \code{RRNG}
object. To get hit information within a \code{RRNG}, use
\code{\link{rngPOS}}.
}
\seealso{
Other ranging functions: 
\code{\link{rangeCount}()},
\code{\link{rangePOS}()},
\code{\link{rngPOS}()}
}
\concept{ranging functions}
