% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_spec.R
\name{rngCount}
\alias{rngCount}
\title{Count Hits in RRNG}
\usage{
rngCount(pos, rng, simplify = FALSE)
}
\arguments{
\item{pos}{data.frame. The \code{POS} or \code{ATO} to be ranged}

\item{rng}{data.frame. The \code{RRNG} ranges to apply
#' @param simplify logical. Whether to simplify the counts by ion name}
}
\value{
A data.frame containing the name of each range, the number of counts
and fraction of the total ranged counts.
}
\description{
\code{rngCount} counts the number of hits for each entry within a \code{RRNG}
object. To get hit information within a \code{RRNG}, use
\code{\link{rngPOS}}.
}
\seealso{
Other ranging functions: 
\code{\link{rangeCount}()},
\code{\link{rangePOS}()},
\code{\link{rngPOS}()}
}
\concept{ranging functions}
