% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_spec.R
\name{rngPOS}
\alias{rngPOS}
\title{Extract Hits in RRNG}
\usage{
rngPOS(pos, rng)
}
\arguments{
\item{pos}{data.frame. The \code{POS} or \code{ATO} from which to extract
hits}

\item{rng}{data.frame. The \code{RRNG} ranges to apply}
}
\value{
A data.frame of the same structure as \code{pos}, with an appended
  column called "mark" that carries the name of the ion
}
\description{
\code{rngPOS} extracts the rows of a \code{POS} or \code{ATO} object whose
mass is within the ranges of the provided \code{RRNG}. To count the hits
within a \code{RRNG}, use \code{\link{rngCount}}
}
\seealso{
Other ranging functions: 
\code{\link{rangeCount}()},
\code{\link{rangePOS}()},
\code{\link{rngCount}()}
}
\concept{ranging functions}
