% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-spat.R
\name{rownames.pp3}
\alias{rownames.pp3}
\title{Extends \code{\link[base]{rownames}} to \code{\link[spatstat]{pp3}}.}
\usage{
rownames.pp3(pat)
}
\description{
Extends \code{\link[base]{rownames}} to \code{\link[spatstat]{pp3}}.
}
