% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_sim.R
\name{rpoint3}
\alias{rpoint3}
\title{Generate N Random Points}
\usage{
rpoint3(
  n,
  f,
  fmax = 1,
  win = box3(),
  ...,
  giveup = 1000,
  verbose = FALSE,
  nsim = 1,
  drop = TRUE
)
}
\description{
\code{rpoint3} extends \code{\link[spatstat.geom]{rpoint}} to
\code{\link[spatstat.geom]{pp3}}.
}
\seealso{
\code{\link[spatstat.geom]{rpoint}}

Other spatstat extensions: 
\code{\link{G3cross}()},
\code{\link{G3multi}()},
\code{\link{K3scaled}()},
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
