% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{sample.pp3}
\alias{sample.pp3}
\title{Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{pp3}}.}
\usage{
sample.pp3(X, size)
}
\arguments{
\item{X}{A \code{pp3}. The point pattern from which to sample.}

\item{size}{A numeric. The number of points to sample.}
}
\value{
A \code{pp3}. The sampled point pattern.
}
\description{
Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{pp3}}.
}
\seealso{
\code{\link[base]{sample}}
}
