% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{sample.pp3}
\alias{sample.pp3}
\title{Sample A 3D Point Pattern}
\usage{
sample.pp3(X, size)
}
\arguments{
\item{X}{A \code{pp3}. The point pattern from which to sample.}

\item{size}{A numeric. The number of points to sample.}
}
\value{
A \code{pp3}. The sampled point pattern.
}
\description{
Sample A 3D Point Pattern
}
\seealso{
\code{\link{sample.ppp}}, \code{\link[base]{sample}}

Other spatstat extensions: 
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
