% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-extend.R
\name{sample.ppp}
\alias{sample.ppp}
\title{Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{ppp}}.}
\usage{
sample.ppp(X, size)
}
\description{
Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{ppp}}.
}
