% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{sample.ppp}
\alias{sample.ppp}
\title{Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{ppp}}.}
\usage{
sample.ppp(X, size)
}
\arguments{
\item{X}{A \code{ppp}. The point pattern from which to sample.}

\item{size}{A numeric. The number of points to sample.}
}
\value{
A \code{ppp}. The sampled point pattern.
}
\description{
Extends \code{\link[base]{sample}} to handle \code{\link[spatstat]{ppp}}.
}
\seealso{
\code{\link[base]{sample}}

Other spatstat extensions: 
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{shift.pp3}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
