% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{sample.ppp}
\alias{sample.ppp}
\title{Sample a Planar Point Pattern}
\usage{
sample.ppp(X, size)
}
\arguments{
\item{X}{A \code{ppp}. The point pattern from which to sample.}

\item{size}{A numeric. The number of points to sample.}
}
\value{
A \code{ppp}. The sampled point pattern.
}
\description{
Sample a Planar Point Pattern
}
\seealso{
\code{\link{sample.pp3}}, \code{\link[base]{sample}}

Other spatstat extensions: 
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
