% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-multi.R
\name{saxeyPlot}
\alias{saxeyPlot}
\title{Creates a 2D correlation histogram of mass spectra and, optionally, plots it.}
\usage{
saxeyPlot(ind, ato, begin, end, res = 0.25, plot.it = T)
}
\description{
\code{saxeyPlot}
}
