% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt-extend.R
\name{shift.pp3}
\alias{shift.pp3}
\title{Extends \code{\link[spatstat]{shift}} to \code{\link[spatstat]{pp3}}.}
\usage{
shift.pp3(X, vec = c(0, 0, 0), ..., origin = NULL)
}
\description{
Extends \code{\link[spatstat]{shift}} to \code{\link[spatstat]{pp3}}.
}
