% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{shift.pp3}
\alias{shift.pp3}
\title{Apply Vector Translation}
\usage{
\method{shift}{pp3}(X, vec = c(0, 0, 0), ..., origin = NULL)
}
\arguments{
\item{X}{Point pattern(object of class"pp3")}

\item{vec}{Numeric. A vector of length 3 specifying the translation}

\item{origin}{Location that will be shifted to the origin. Either a vector of
length 3 specifying the new origin or one of the character strings
"midpoint" or "bottomleft" (will be extended)}
}
\description{
Applies a vector shift to a 3D point pattern
}
\seealso{
\code{\link[spatstat]{shift}}

Other spatstat extensions: 
\code{\link{G3cross}()},
\code{\link{G3multi}()},
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
