% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{studpermu.test.pp3}
\alias{studpermu.test.pp3}
\title{Studentised Permutation Test}
\usage{
\method{studpermu.test}{pp3}(
  X,
  formula,
  summaryfunction = K3est,
  ...,
  nperm = 999,
  use.Tbar = FALSE,
  rinterval = NULL,
  minpoints = 20,
  rmax = NULL,
  nrval = 128
)
}
\arguments{
\item{X}{A hyperframe containing at least the point patterns and groups}

\item{formula}{Formula describing the grouping. The left side of the formula
identifies which column of X contains the point patterns. The right side
identifies the grouping factor}

\item{summaryfunction}{Summary function applicable to pp3. Defaults to
\code{link[spatstat]{K3est}}}

\item{...}{Additional arguments passed to summaryfunction}

\item{nperm}{Number of random permutations for the test; defaults to 999}

\item{use.Tbar}{Logical value indicating choice of test statistic. If TRUE,
use the alternative test statistic, which is appropriate for summary
functions with roughly constant variance, such as K(r)/r or L(r). Defaults
to FALSE}

\item{rinterval}{Numeric of length 2. Experimental}

\item{minpoints}{Minimum permissible number of points in a point pattern for
inclusion in the test calculation}
}
\description{
Perform a studentised permutation test for a difference between groups of
point patterns
}
\references{
Hahn, U. (2012) A studentized permutation test for the comparison
  of spatial point patterns.
  Journal of the American Statistical Association 107 (498), 754-764.
}
\seealso{
\code{\link[spatstat]{studpermu.test}}, \code{link[spatstat]{K3est}}

Other spatstat extensions: 
\code{\link{Tstat.pp3}()},
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
