% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rTEM_functions.R
\name{summary_functions}
\alias{summary_functions}
\title{function to calculate all summary functions on point pattern}
\usage{
summary_functions(
  funcs = c("K", "G", "F", "GXGH"),
  pattern,
  ...,
  dopant_formula,
  host_formula,
  K_cor = "trans",
  G_cor = "km",
  F_cor = "km",
  GXGH_cor = "km",
  GXHG_cor = "km"
)
}
\arguments{
\item{funcs}{vector of summary functions to calculate}

\item{pattern}{point pattern of type ppp or pp3}

\item{...}{maxKr, nKr, maxGr, nGr, maxGXGHr, maxGXHGr, nGXr}

\item{dopant_formula}{mark of points from which to calculate summary functions}

\item{host_formula}{marks of points to not use in summary functions (except cross type functions)}

\item{K_cor}{edge correction(s) to be used for K function}

\item{G_cor}{edge correction(s) to be used for G function}

\item{F_cor}{edge correction(s) to be used for F function}

\item{GXGH_cor}{edge correction(s) to be used for GXGH function}

\item{GXHG_cor}{edge correction(s) to be used for GXHG function}
}
\description{
Cacluate that summary functions included in `funcs` on `pattern`, a point pattern
of class ppp or pp3.
}
