% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{superimpose.pp3}
\alias{superimpose.pp3}
\title{Extends \code{\link[spatstat]{superimpose}} to \code{\link[spatstat]{pp3}}.}
\usage{
\method{superimpose}{pp3}(..., W = NULL, check = F)
}
\description{
\code{superimpose.pp3}
}
\seealso{
\code{\link[spatstat]{superimpose}}

Other spatstat extensions: 
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()}
}
\concept{spatstat extensions}
