% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_extend.R
\name{Tstat.pp3}
\alias{Tstat.pp3}
\title{Extends Tstat to pp3}
\usage{
Tstat.pp3(
  X,
  ...,
  rmax = NULL,
  nrval = 128,
  correction = "border",
  ratio = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{correction}{One of "none" or "isotropic." "translation" correction is
planned but not yet implemented.}
}
\description{
Tstat.pp3 extends the third-order summary statistic
\code{\link[spatstat]{Tstat}} to pp3
}
\references{
Schladitz, K. and Baddeley, A. (2000) A third order point process
characteristic. Scandinavian Journal of Statistics 27 (2000) 657-671.
}
\seealso{
\code{\link[spatstat]{Tstat}}

Other spatstat extensions: 
\code{\link{bdist.points}()},
\code{\link{marktable.pp3}()},
\code{\link{marktable}()},
\code{\link{quadratcount.pp3}()},
\code{\link{quadrats.pp3}()},
\code{\link{rPoissonCluster3}()},
\code{\link{rjitter.pp3}()},
\code{\link{rjitter.ppp}()},
\code{\link{rjitter}()},
\code{\link{rpoint3}()},
\code{\link{sample.pp3}()},
\code{\link{sample.ppp}()},
\code{\link{shift.pp3}()},
\code{\link{studpermu.test.pp3}()},
\code{\link{studpermu.test}()},
\code{\link{superimpose.pp3}()}
}
\concept{spatstat extensions}
