% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapt_file.R
\name{writePOS}
\alias{writePOS}
\title{Write a POS File}
\usage{
writePOS(pos, filepath)
}
\arguments{
\item{pos}{A data.frame with columns of x, y, z, and mass, following the
structure of the \code{data.frame} returned by \code{\link{readPOS}}.}

\item{filepath}{A string. The file path to the POS file.}
}
\description{
\code{writePOS} writes a POS \code{data.frame} into a POS file
}
\references{
Larson \emph{et al.},
\emph{Local Electrode Atom Probe Tomography: A User's Guide} (2009)
}
\seealso{
\code{\link{readPOS}}
}
